% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_negbinom.R
\name{d_negbinom}
\alias{d_negbinom}
\title{Fonction de densité de la loi binomiale négative}
\usage{
d_negbinom(k, r, p = (1/(1 + beta)), beta = ((1 - p)/p),
  nb_tries = F)
}
\arguments{
\item{k}{Nombre de échecs (ou succès) visé.}

\item{r}{Nombre de succès visé.}

\item{p}{Probabilité de succès.}

\item{beta}{Deuxième paramétrisation de la loi binomiale négative tel que beta = (1 - p) / p.}

\item{nb_tries}{Stipule si la binomiale est une du nombre d'essais pour le r-ème succès, ou du nombre d'échecs pour le r-ème succès (par défaut).}
}
\description{
Fonction de densité de la loi binomiale négative
}
