% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaMixture.R
\name{dgammaMixture}
\alias{dgammaMixture}
\title{Mixture of Gamma distributions}
\usage{
dgammaMixture(x, shape, rate = 1/scale, scale = 1/rate, TOL = 1e-06)
}
\arguments{
\item{shape}{vector of shape parameters \eqn{\alpha_i}{alpha_i} (i = 1, 2). Vector must be of length 2. Parameters must be positive.}

\item{rate}{vector of rate parameters \eqn{\beta_i}{beta_i} (i = 1, 2). Vector must be of length 2.}

\item{scale}{alternative parametrisation of rate parameters. \eqn{\frac{1}{\beta}}{scale = 1 / rate}.}

\item{TOL}{Error for the finding the upper bound limit of the summation. 1E-6 by default.
<%=ifelse(exists("x") && x, "@param x quantile. Doit être strictement supérieur à 0.", "") %>
<%=ifelse(exists("q") && q, "@param q quantile.", "") %>
<%=ifelse(exists("lower.tail") && lower.tail, "@param lower.tail logical; if \\code{TRUE} (défaut) probabilities are Pr(M <= k), otherwise, Pr(M > k).", "") %>
<%=ifelse(exists("kappa") && kappa, "@param kappa probability.", "") %>
<%=ifelse(exists("d") && d, "@param d cut-off value.", "") %>}
}
\description{
Probability mass function, cumulative probability mass function,
 expected value, truncated expected value, and Tail Value-at-Risk (TVaR) of a
 mixture of Gamma distributions with shape parameters \eqn{\alpha_i}{alpha_i}
 and rate parameters \eqn{\beta_i}{beta_i} (i = 1, 2).
}
\examples{
pgammaMixture(q = 15, shape = c(3, 5), rate = c(1/7, 1/4))
pgammaMixture(q = 15, shape = c(3, 5), scale = c(7, 4))
pgammaMixture(q = 15, shape = c(3, 5), rate = c(1/7, 1/4), lower.tail = FALSE)

dgammaMixture(x = 15, shape = c(3, 5), scale = c(7, 4))


}
\concept{Gamma mixture}
