% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dIG.R
\name{dIG}
\alias{dIG}
\title{Envelope de la densité de la loi inverse gaussienne}
\usage{
dIG(x, mu, beta = dispersion * mu^2, dispersion = beta/mu^2)
}
\arguments{
\item{x}{x}

\item{mu}{mu parameter}

\item{beta}{beta, = dispersion * mu^2}

\item{dispersion}{beta, = beta / mu^2}
}
\description{
Envelope de la densité de la loi inverse gaussienne
}
\details{
Wrapper pour avoir une notation cohérente avec le reste des fonctions IG, mais utilise la fonction de densité du package ACTUAR.
}
