% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributacalcul-package.R
\docType{package}
\name{Distributacalcul-package}
\alias{Distributacalcul}
\alias{Distributacalcul-package}
\title{Distributacalcul: Probability Distribution Functions}
\description{
The Distributacalcul provides three general types of functions :
 \enumerate{
  \item  Functions related to continuous, discrete, and compound probability distributions.
  \item  Functions related to bivariate copulas.
  \item  Shiny applications for visualising these distributions and copulas.
 }
}
\section{Probability Distribution Functions}{

These include functions for calculating various moments, risk measures,
and distribution functions.
}

\section{Copula Functions}{

These include functions for calculating distribution functions and
simulating observations from the copulas.
}

\section{Shiny Applications}{

These include functions to plot the distributions as well as to calculate
the moments and risk measures.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://alec42.github.io/Distributacalcul_Package/}
  \item Report bugs at \url{https://github.com/alec42/Distributacalcul_Package/issues}
}

}
\author{
\strong{Maintainer}: Alec James van Rassel \email{alec.van-rassel.1@ulaval.ca} [copyright holder]

Authors:
\itemize{
  \item Gabriel Crépeault-Cauchon [conceptor]
}

Other contributors:
\itemize{
  \item Étienne Marceau [teacher, scientific advisor]
  \item Hélène Cossette [teacher, scientific advisor]
  \item Laboratoire Act & Risk [funder, supporting host]
  \item École d'actuariat de l'Université Laval [funder, host institution, university place]
  \item Natural Sciences and Engineering Research Council of Canada [funder]
  \item Marc-André Devost [conceptor]
}

}
\keyword{internal}
