% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_shiny.R
\name{Distributacalcul_vis}
\alias{Distributacalcul_vis}
\title{Interactive distribution visualisation}
\usage{
Distributacalcul_vis(law)
}
\arguments{
\item{law}{Distribution to visualise, one of ...}
}
\description{
Interactive distribution visualisation
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
   Distributacalcul_vis("norm")
}

}
