% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionPlotBoxCopulas.R
\name{distributionPlotBoxCopulasUI}
\alias{distributionPlotBoxCopulasUI}
\title{Interactive moments visualization (UI side) for copulas}
\usage{
distributionPlotBoxCopulasUI(id)
}
\arguments{
\item{id}{id of module}
}
\value{
UI function for the moments module of the Copula Shiny
 function. Should not be run directly.
}
\description{
Interactive moments visualization (UI side) for copulas
}
\keyword{internal}
