% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlogarithmic.R
\name{dlogarithmic}
\alias{dlogarithmic}
\title{Probability Mass Function of the Logarithmic Distribution}
\usage{
dlogarithmic(x, prob)
}
\arguments{
\item{x}{quantile.}

\item{prob}{probability parameter \eqn{\gamma}{gamma}.}
}
\value{
Function :
 \itemize{
 \item{\code{\link{dlogarithmic}}}{ gives the probability mass function (PMF).}
 \item{\code{\link{plogarithmic}}}{ gives the cumulative mass function (CMF).}
 \item{\code{\link{VaR_logarithmic}}}{ gives the Value-at-Risk (VaR), a.k.a quantile, function.}
 \item{\code{\link{MGF_logarithmic}}}{ gives the moment generating function (MGF).}
 \item{\code{\link{PGF_logarithmic}}}{ gives the probability generating function (PGF).}
 \item{\code{\link{E_logarithmic}}}{ gives the expected value.}
 \item{\code{\link{V_logarithmic}}}{ gives the variance.}
 }
 Invalid parameter values will return an error detailing which parameter is problematic.
 Returned values of the VaR are approximations.
}
\description{
Probability mass function of the Logarithmic distribution with
 probability parameter \eqn{\gamma}{gam}.
}
\details{
The Logarithmic distribution with probability parameter \eqn{\gamma}{gam}
 has probability mass function
 \deqn{Pr(X = k) = \frac{-\gamma^{k}}{\ln(1 - \gamma)k}}{-gam^k / (ln(1 - k) k)},
 for \eqn{k = 0, 1, 2, \dots}{k = 0, 1, 2, ...}, and \eqn{\gamma  \in (0, 1)}{0 < gam < 1}]
}
\examples{

# With rate parameter
dlogarithmic(x = 3, prob = 0.2)

}
\seealso{
Other Logarithmic Distribution: 
\code{\link{E_logarithmic}()},
\code{\link{MGF_logarithmic}()},
\code{\link{PGF_logarithmic}()},
\code{\link{V_logarithmic}()},
\code{\link{VaR_logarithmic}()},
\code{\link{plogarithmic}()}
}
\concept{Logarithmic Distribution}
