% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFGEN.R
\name{.pFGEN}
\alias{.pFGEN}
\title{Cumulative density function of the generalized Fisher-Snedecor distribution}
\usage{
.pFGEN(x, lambda, alpha, tau)
}
\arguments{
\item{x}{x}

\item{lambda}{lambda, > 0}

\item{alpha}{alpha, >0}

\item{tau}{tau, >0}
}
\description{
Cumulative density function of the generalized Fisher-Snedecor distribution
}
