% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_erlang.R
\name{.TVaR_erlang}
\alias{.TVaR_erlang}
\title{Tail Value-at-Risk of the Erlang Distribution}
\usage{
.TVaR_erlang(kappa, shape, scale, rate = 1/scale, vark)
}
\arguments{
\item{shape}{shape parameter \eqn{n}, must be positive integer.}

\item{scale}{\eqn{\beta}{beta} is the scale parameter, must be positive.}

\item{rate}{alternative parameterization to scale parameter, rate = 1 / scale.}
}
\description{
Tail Value-at-Risk of the Erlang distribution with shape
 parameter \eqn{n} and rate parameter \eqn{\beta}{b}.
}
\details{
Cette formule nécessite la formule de la VaR_erlang (déjà installée avec le package tvarPackage)
}
\seealso{
[derlang()] for the probability mass function,
 [perlang()] for the cumulative probability mass function,
 [V_erlang()] for the variance,
 [Mexcess_erlang()] for the Mean Excess Loss,
 [SL_erlang()] for the stop-loss,
 [Elim_erlang()] for the limited expected value,
 [Etronq_erlang()] for the truncated mean,
 [E_erlang()] for the expected value, and
 [kthmoment_erlang()] for the \eqn{k}th moment.
}
