% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisgen.R
\name{dpoisGen}
\alias{dpoisGen}
\title{Generalized Poisson-mixture}
\usage{
dpoisGen(k, shape1, shape2)
}
\arguments{
\item{k}{probability.}

\item{shape1}{shape parameter \eqn{\lambda}{lambda}, must be positive.}

\item{shape2}{shape parameter \eqn{\tau}{tau}, where \eqn{\tau \in [0, )}{0 <= tau < 1}}

\item{q}{quantile, \eqn{k \in \mathcal{N}}{k is a positive integer}.}
}
\description{
Probability mass function, cumulative probability mass function,
 expected value and variance of the Generalized Poisson-mixture with
 shape parameters \eqn{\lambda}{lambda} and \eqn{\tau}{tau}.
}
\examples{
dpoisGen(k = 2, shape1 = 0.25, shape2 = 0.5)
ppoisGen(q = 3, shape1 = 0.25, shape2 = 0.5)
E_poisGen(shape1 = 0.25, shape2 = 0.5)
V_poisGen(shape1 = 0.25, shape2 = 0.5)

}
\concept{Mixture distributions}
