% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunifD.R
\name{dunifD}
\alias{dunifD}
\title{Probability mass function of the (discrete) Uniform distribution}
\usage{
dunifD(x, min = 0, max = 1)
}
\arguments{
\item{x}{quantile. By definition, it has no impact on the uniform distribution. Set to 1 by default.}

\item{min, max}{lower and upper limits of the distribution. Must be finite.}
}
\description{
Probability mass function of the (discrete) Uniform
 distribution with min \eqn{a} and max \eqn{b}.
}
\examples{

# With scale parameter
dunifD(x = 2, min = 2, max = 5)

}
