% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_BINCOMP.R
\name{E_BINCOMP}
\alias{E_BINCOMP}
\title{Espérance d'une loi binomiale composée}
\usage{
E_BINCOMP(size, prob, shape, rate, distr_severity = "Gamma")
}
\arguments{
\item{size}{n pour la binomiale}

\item{prob}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{distr_severity}{choix de distribution de sévérité. Gamma (par défaut) ou Lognormale}
}
\description{
Espérance d'une loi binomiale composée
}
\details{
Cette formule utilise aussi un choix de 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
\examples{
E_BINCOMP(size = 1, prob = 0.2, shape = log(1000) - 0.405, rate = 0.9^2,
          distr_severity = "Lognormale")

}
