% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_binom.R
\name{E_binom}
\alias{E_binom}
\title{Expected value of the Binomial distribution}
\usage{
E_binom(size, prob)
}
\arguments{
\item{size}{Number of trials (0 or more).}

\item{prob}{Probability of success on each trial.}
}
\description{
Expected value of the Binomial distribution with size
 \eqn{n}{n} and probability of success \eqn{p}{p}.
}
\details{

}
\examples{

E_binom(size = 3, prob = 0.5)

}
\seealso{
Other Binomial Distribution: 
\code{\link{Etronq_binom}()},
\code{\link{MGF_binom}()},
\code{\link{V_binom}()}
}
\concept{Binomial Distribution}
