% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_BNCOMP.R
\name{E_BNCOMP}
\alias{E_BNCOMP}
\title{Espérance d'une loi binomiale négative composée}
\usage{
E_BNCOMP(r, q, shape, rate, distr_severity = "Gamma")
}
\arguments{
\item{r}{r pour la binomiale négative}

\item{q}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{distr_severity}{choix de distribution de sévérité. Gamma (par défaut) ou Lognormale}
}
\description{
Espérance d'une loi binomiale négative composée
}
\details{
Cette formule utilise aussi un choix de 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
