% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_hyper.R
\name{E_hyper}
\alias{E_hyper}
\title{Espérance d'une loi hypergéométrique}
\usage{
E_hyper(N, m, n)
}
\arguments{
\item{N}{nombre total}

\item{m}{nombre de type 1}

\item{n}{taille de l'échantillon}
}
\description{
Espérance d'une loi hypergéométrique
}
