% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_IG.R
\name{E_IG}
\alias{E_IG}
\title{Espérance d'une loi inverse gaussienne}
\usage{
E_IG(mu, beta = dispersion * mu^2, dispersion = beta/mu^2)
}
\arguments{
\item{mu}{mu}

\item{beta}{beta = dispersion * mu^2}

\item{dispersion}{dispersion = beta / mu^2}
}
\description{
Espérance d'une loi inverse gaussienne
}
