% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_logarithmique.R
\name{E_logarithmique}
\alias{E_logarithmique}
\title{Espérance d'une loi logarithmique}
\usage{
E_logarithmique(gam)
}
\arguments{
\item{gam}{paramètre gamma}
}
\description{
Espérance d'une loi logarithmique
}
