% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_nbinom.R
\name{E_nbinom}
\alias{E_nbinom}
\title{Espérance d'une loi binomiale négative
Calcul l'espérance d'une binomiale négative}
\usage{
E_nbinom(r, q)
}
\arguments{
\item{r}{nombre d'essais}

\item{q}{probabilité d'un succès indépendant}
}
\description{
Espérance d'une loi binomiale négative
Calcul l'espérance d'une binomiale négative
}
