% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_norm.R
\name{E_norm}
\alias{E_norm}
\title{Expected value of the Normal distribution}
\usage{
E_norm(mean, sd)
}
\arguments{
\item{mean}{mean \eqn{\mu}{mu}.}

\item{sd}{standard deviation \eqn{\sigma}{sigma}}
}
\description{
Expected value of the Normal distribution with mean
 \eqn{\mu}{mu} and variance \eqn{\sigma}{sigma}.
}
\details{
The Normal distribution with mean \eqn{\mu}{mu} and standard deviation
  \eqn{\sigma}{sigma} has density:
  \deqn{\frac{1}{\sqrt{2\pi}\sigma}\textrm{e}^{-\frac{1}{2}\left(\frac{x - \mu}{\sigma}\right)^2}}{f(x) = e^(-(1/2) ((x - mu)/sigma)^2) / ((2 pi)^(1/2) sigma}
  for \eqn{x \in \mathcal{R}}{x real}, \eqn{\mu \in \mathcal{R}, \sigma > 0}{mu real, sigma > 0}.
}
\examples{

E_norm(mean = 3, sd = 5)

}
\seealso{
Other Normal distribution: 
\code{\link{Elim_norm}()},
\code{\link{Etronq_norm}()},
\code{\link{MGF_norm}()},
\code{\link{Mexcess_norm}()},
\code{\link{SL_norm}()},
\code{\link{TVaR_norm}()},
\code{\link{V_norm}()},
\code{\link{VaR_norm}()}
}
\concept{Normal distribution}
