% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_PCOMP.R
\name{E_PCOMP}
\alias{E_PCOMP}
\title{Expected value of a compound Poisson distribution}
\usage{
E_PCOMP(lambda, shape, rate, distr_severity = "Gamma")
}
\arguments{
\item{lambda}{paramètre lambda pour la poisson}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{distr_severity}{choix de distribution de sévérité. Gamma (par défaut) ou Lognormale}
}
\description{
Expected value of a compound Poisson distribution
}
\details{
Cette formule utilise aussi un choix de 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
