% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisgen.R
\name{E_poisGenComp}
\alias{E_poisGenComp}
\title{Compound Generalized Poisson-mixture}
\usage{
E_poisGenComp(shape1, shape2, paramSev, distribution = "Gamma")
}
\arguments{
\item{paramSev}{vector of severity distribution parameters.}

\item{distribution}{severity distribution choice. Valid arguments include "Gamma".}

\item{upper}{Upper bound for the summation of the cumulative density function approximation. In the case of the TVaR, it is a vector to optimize (in order) the VaR and the TVaR.}
}
\description{
Expected value and variance of the Compound Generalized Poisson-mixture
 with shape parameters \eqn{\lambda}{lambda} and \eqn{\tau}{tau}.
 Additionnaly, approximations of the cumulative density function (CDF), Value-at-Risk (VaR),
 Tail Value-at-Risk (TVaR) are included.
}
\examples{
E_poisGenComp(shape1 = 0.25, shape2 = 0.5, paramSev = c(2, 1/1000))
V_poisGenComp(shape1 = 0.25, shape2 = 0.5, paramSev = c(2, 1/1000))
Etronq_poisGenComp(d = 10193.43, shape1 = 0.25, shape2 = 0.5, paramSev = c(2, 1/1000))
ppoisGenComp(q = 10193.43, shape1 = 0.25, shape2 = 0.50, paramSev = c(2, 1/1000))
VaR_poisGenComp(kappa = 0.99, shape1 = 0.25, shape2 = 0.50, paramSev = c(2, 1/1000), upper = 1e5)
Etronq_poisGenComp(VaR_k = 10193.43, shape1 = 0.25, shape2 = 0.50, paramSev = c(2, 1/1000))

}
\concept{Compound mixture distributions}
