% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_unif.R
\name{E_unif}
\alias{E_unif}
\title{Espérance d'une loi uniforme
Calcul l'espérance d'une uniforme}
\usage{
E_unif(a, b)
}
\arguments{
\item{a}{minimum}

\item{b}{maximum}
}
\description{
Espérance d'une loi uniforme
Calcul l'espérance d'une uniforme
}
