% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_unifD.R
\name{E_unifD}
\alias{E_unifD}
\title{Expected value of the (discrete) Uniform distribution}
\usage{
E_unifD(min = 0, max = 1)
}
\arguments{
\item{min, max}{lower and upper limits of the distribution. Must be finite.}
}
\description{
Expected value of the (discrete) Uniform distribution
 with min \eqn{a} and max \eqn{b}.
}
\examples{

# With scale parameter
E_unifD(min = 2, max = 5)

}
