% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_beta.R
\name{Elim_beta}
\alias{Elim_beta}
\title{Espérance limitée d'une loi Beta}
\usage{
Elim_beta(d, a, b)
}
\arguments{
\item{d}{déductible (doit être entre 0 et 1)}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Espérance limitée d'une loi Beta
}
