% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_erlang.R
\name{Elim_erlang}
\alias{Elim_erlang}
\title{Espérance limitée d'une loi Erlang}
\usage{
Elim_erlang(d, n, b)
}
\arguments{
\item{d}{déductible}

\item{n}{n}

\item{b}{beta}
}
\description{
Espérance limitée d'une loi Erlang
}
