% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_exp.R
\name{Elim_exp}
\alias{Elim_exp}
\title{Espérance limitée d'une loi exponentielle}
\usage{
Elim_exp(d, b)
}
\arguments{
\item{d}{montant d de la limite}

\item{b}{beta}
}
\description{
Espérance limitée d'une loi exponentielle
}
