% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_gamma.R
\name{Elim_gamma}
\alias{Elim_gamma}
\title{Espérance limitée d'une loi gamma}
\usage{
Elim_gamma(d, a, b)
}
\arguments{
\item{d}{déductible}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Espérance limitée d'une loi gamma
}
