% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_IG.R
\name{Elim_IG}
\alias{Elim_IG}
\title{Espérance limitée d'une loi inverse gaussienne}
\usage{
Elim_IG(d, mu, beta = dispersion * mu^2, dispersion = beta/mu^2)
}
\arguments{
\item{d}{montant d de la limite}

\item{mu}{mu}

\item{beta}{beta, = dispersion * mu^2}

\item{dispersion}{beta, = beta / mu^2}
}
\description{
Espérance limitée d'une loi inverse gaussienne
}
