% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_llogis.R
\name{Elim_llogis}
\alias{Elim_llogis}
\title{Espérance limitée d'une loi Log-logistique}
\usage{
Elim_llogis(d, lam, tau)
}
\arguments{
\item{d}{déductible}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Espérance limitée d'une loi Log-logistique
}
