% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_lnorm.R
\name{Elim_lnorm}
\alias{Elim_lnorm}
\title{Espérance limitée d'une loi lognormale}
\usage{
Elim_lnorm(d, mu, sig)
}
\arguments{
\item{d}{déductible}

\item{mu}{mu}

\item{sig}{sigma}
}
\description{
Espérance limitée d'une loi lognormale
}
