% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_norm.R
\name{Elim_norm}
\alias{Elim_norm}
\title{Espérance limitée de la loi normale}
\usage{
Elim_norm(d, mu = 0, sig = 1)
}
\arguments{
\item{d}{déductible}

\item{mu}{moyenne mu}

\item{sig}{sigma}
}
\description{
Espérance limitée de la loi normale
}
