% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_unif.R
\name{Elim_unif}
\alias{Elim_unif}
\title{Espérance limités d'une loi uniforme}
\usage{
Elim_unif(d, a, b)
}
\arguments{
\item{a}{minimum}

\item{b}{maximum}
}
\description{
Espérance limités d'une loi uniforme
}
