% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_weibull.R
\name{Elim_weibull}
\alias{Elim_weibull}
\title{Espérance limitée d'une loi Weibull}
\usage{
Elim_weibull(d, tau, beta)
}
\arguments{
\item{d}{déductible}

\item{tau}{tau}

\item{beta}{beta}
}
\description{
Espérance limitée d'une loi Weibull
}
