% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_beta.R
\name{Etronq_beta}
\alias{Etronq_beta}
\title{Espérance tronquée d'une loi Beta}
\usage{
Etronq_beta(d, a, b)
}
\arguments{
\item{d}{déductible}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Espérance tronquée d'une loi Beta
}
