% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_burr.R
\name{Etronq_burr}
\alias{Etronq_burr}
\title{Espérance tronquée d'une loi Burr}
\usage{
Etronq_burr(d, alpha, lam, tau)
}
\arguments{
\item{d}{Deductible}

\item{alpha}{alpha}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Espérance tronquée d'une loi Burr
}
