% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_gamma.R
\name{Etronq_gamma}
\alias{Etronq_gamma}
\title{Espérance tronquée d'une loi Gamma}
\usage{
Etronq_gamma(d, a, b)
}
\arguments{
\item{d}{Deductible}

\item{a}{alpha}

\item{b}{beta (ou lambda dans certains livres)}
}
\description{
Espérance tronquée d'une loi Gamma
}
