% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_IG.R
\name{Etronq_IG}
\alias{Etronq_IG}
\title{Espérance tronquée d'une loi inverse gaussienne}
\usage{
Etronq_IG(d, mu, beta = dispersion * mu^2, dispersion = beta/mu^2)
}
\arguments{
\item{d}{valeur où tronquer}

\item{mu}{mu}

\item{beta}{beta = dispersion * mu^2}

\item{dispersion}{dispersion = beta / mu^2}
}
\description{
Espérance tronquée d'une loi inverse gaussienne
}
