% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_llogstique.R
\name{Etronq_llogis}
\alias{Etronq_llogis}
\title{Espérance tronquée d'une loi Log-logistique}
\usage{
Etronq_llogis(d, lam, tau)
}
\arguments{
\item{d}{Deductible}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Espérance tronquée d'une loi Log-logistique
}
