% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_lnorm.R
\name{Etronq_lnorm}
\alias{Etronq_lnorm}
\title{Espérance tronquée d'une loi lognormale}
\usage{
Etronq_lnorm(d, mu, sig)
}
\arguments{
\item{d}{déductible}

\item{mu}{mu}

\item{sig}{sigma}
}
\description{
Espérance tronquée d'une loi lognormale
}
