% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_norm.R
\name{Etronq_norm}
\alias{Etronq_norm}
\title{Espérance tronquée de la loi normale}
\usage{
Etronq_norm(d, mu = 0, sig = 1)
}
\arguments{
\item{d}{déductible}

\item{mu}{moyenne mu}

\item{sig}{sigma}
}
\description{
Espérance tronquée de la loi normale
}
