% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_pareto.R
\name{Etronq_pareto}
\alias{Etronq_pareto}
\title{Espérance tronquée d'une loi Pareto}
\usage{
Etronq_pareto(d, alph, lam)
}
\arguments{
\item{alph}{alpha}

\item{lam}{lambda}
}
\description{
Espérance tronquée d'une loi Pareto
}
