% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_unif.R
\name{Etronq_unif}
\alias{Etronq_unif}
\title{Espérance tronquée d'une loi uniforme}
\usage{
Etronq_unif(d, a, b)
}
\arguments{
\item{a}{minimum}

\item{b}{maximum}
}
\description{
Espérance tronquée d'une loi uniforme
}
