% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_weibull.R
\name{Etronq_weibull}
\alias{Etronq_weibull}
\title{Espérance tronquée d'une loi Weibull}
\usage{
Etronq_weibull(d, tau, beta)
}
\arguments{
\item{d}{valeur où tronquer}

\item{tau}{tau}

\item{beta}{beta}
}
\description{
Espérance tronquée d'une loi Weibull
}
