% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsBox.R
\name{functionsBoxUI}
\alias{functionsBoxUI}
\title{Interactive moments visualization (UI side)}
\usage{
functionsBoxUI(id)
}
\arguments{
\item{id}{id of module}
}
\value{
UI function for the moments module.
 Should not be run directly.
}
\description{
Interactive moments visualization (UI side)
}
\keyword{internal}
