% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgompertz.R
\name{hgompertz}
\alias{hgompertz}
\title{Force de mortalité de la loi Gompertz}
\usage{
hgompertz(t = 0, x = 0, beta, gam, Tx = T)
}
\arguments{
\item{t}{durée de vie T_x}

\item{x}{âge ateint}

\item{beta}{paramètre beta}

\item{gam}{paramètre gamma}

\item{Tx}{Paramètre Tx, si on désire avoir la distribution de Tx.}
}
\description{
Force de mortalité de la loi Gompertz
}
