% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_beta.R
\name{kthmoment_beta}
\alias{kthmoment_beta}
\title{k-ème moment de la loi Beta}
\usage{
kthmoment_beta(k, a, b)
}
\arguments{
\item{k}{k-ème moment}

\item{a}{alpha}

\item{b}{beta}
}
\description{
k-ème moment de la loi Beta
}
