% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_burr.R
\name{kthmoment_burr}
\alias{kthmoment_burr}
\title{k-ème moment d'une loi Burr}
\usage{
kthmoment_burr(k, alpha, lam, tau)
}
\arguments{
\item{k}{k-ème moment, -tau < k < alpha * tau}

\item{alpha}{alpha}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
k-ème moment d'une loi Burr
}
