% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_gamma.R
\name{kthmoment_gamma}
\alias{kthmoment_gamma}
\title{k-ème moment de la loi gamma}
\usage{
kthmoment_gamma(k, shape, rate)
}
\arguments{
\item{k}{k-ème moment}

\item{shape}{alpha}

\item{rate}{beta}
}
\description{
k-ème moment de la loi gamma
}
