% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_llogis.R
\name{kthmoment_llogis}
\alias{kthmoment_llogis}
\title{k-ème moment d'une loi log-logistique}
\usage{
kthmoment_llogis(k = 1, lam, tau)
}
\arguments{
\item{k}{k-ème moment}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
k-ème moment d'une loi log-logistique
}
