% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_lnorm.R
\name{kthmoment_lnorm}
\alias{kthmoment_lnorm}
\title{k-ème moment de la loi Lognormale}
\usage{
kthmoment_lnorm(k, mu, sig)
}
\arguments{
\item{k}{k-ème moment}

\item{mu}{mu}

\item{sig}{sigma = sqrt(sigma^2)}
}
\description{
k-ème moment de la loi Lognormale
}
