% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_unif.R
\name{kthmoment_unif}
\alias{kthmoment_unif}
\title{kth moment of the Uniform distribution}
\usage{
kthmoment_unif(k, min, max)
}
\arguments{
\item{min}{minimum}

\item{max}{maximum}
}
\description{
kth moment of the Uniform distribution with shape parameters
 \eqn{a}{a} and \eqn{b}{b}.
}
\examples{

kthmoment_unif(k = 3, min = 4, max = 5)

}
