% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_weibull.R
\name{kthmoment_weibull}
\alias{kthmoment_weibull}
\title{k-ème moment de la loi Weibull}
\usage{
kthmoment_weibull(tau, beta, k = 1)
}
\arguments{
\item{tau}{tau}

\item{beta}{beta}

\item{k}{k-ème moment}
}
\description{
k-ème moment de la loi Weibull
}
