% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_beta.R
\name{Mexcess_beta}
\alias{Mexcess_beta}
\title{Mean Excess-Loss  d'une loi Beta}
\usage{
Mexcess_beta(d, a, b)
}
\arguments{
\item{d}{déductible}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Mean Excess-Loss  d'une loi Beta
}
