% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_erlang.R
\name{Mexcess_erlang}
\alias{Mexcess_erlang}
\title{Mean-Excess loss (excès-moyen) d'une loi erlang}
\usage{
Mexcess_erlang(d, n, b)
}
\arguments{
\item{d}{déductible}

\item{n}{n}

\item{b}{beta}
}
\description{
Mean-Excess loss (excès-moyen) d'une loi erlang
}
