% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_gamma.R
\name{Mexcess_gamma}
\alias{Mexcess_gamma}
\title{Mean-Excess loss d'une loi gamma}
\usage{
Mexcess_gamma(d, a, b)
}
\arguments{
\item{d}{déductible}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Mean-Excess loss d'une loi gamma
}
