% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_IG.R
\name{Mexcess_IG}
\alias{Mexcess_IG}
\title{Mean-Excess loss (excès-moyen) d'une loi inverse gaussienne}
\usage{
Mexcess_IG(d, mu, beta = dispersion * mu^2, dispersion = beta/mu^2)
}
\arguments{
\item{d}{déductible}

\item{mu}{mu}

\item{beta}{beta = dispersion * mu^2}

\item{dispersion}{dispersion = beta / mu^2}
}
\description{
Mean-Excess loss (excès-moyen) d'une loi inverse gaussienne
}
