% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_llogistique.R
\name{Mexcess_llogis}
\alias{Mexcess_llogis}
\title{Mean-Excess loss d'une loi Log-logistique}
\usage{
Mexcess_llogis(d, lam, tau)
}
\arguments{
\item{d}{déductible}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Mean-Excess loss d'une loi Log-logistique
}
