% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_lnorm.R
\name{Mexcess_lnorm}
\alias{Mexcess_lnorm}
\title{Mean Excess-Loss d'une loi lognormale}
\usage{
Mexcess_lnorm(d, mu, sig)
}
\arguments{
\item{d}{déductible}

\item{mu}{mu}

\item{sig}{sigma}
}
\description{
Mean Excess-Loss d'une loi lognormale
}
