% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_norm.R
\name{Mexcess_norm}
\alias{Mexcess_norm}
\title{Mean-Excess Loss de la loi normale}
\usage{
Mexcess_norm(d, mu = 0, sig = 1)
}
\arguments{
\item{d}{déductible}

\item{mu}{moyenne mu}

\item{sig}{sigma}
}
\description{
Mean-Excess Loss de la loi normale
}
