% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_pareto.R
\name{Mexcess_pareto}
\alias{Mexcess_pareto}
\title{Mean Excess-Loss d'une loi Pareto}
\usage{
Mexcess_pareto(d, alph, lam)
}
\arguments{
\item{d}{déductible}

\item{alph}{alpha}

\item{lam}{lambda}
}
\description{
Mean Excess-Loss d'une loi Pareto
}
