% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_unif.R
\name{Mexcess_unif}
\alias{Mexcess_unif}
\title{Mean excess loss d'une loi uniforme}
\usage{
Mexcess_unif(d, a, b)
}
\arguments{
\item{d}{cut-off value}

\item{a}{minimum}

\item{b}{maximum}
}
\description{
Mean excess loss d'une loi uniforme
}
