% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_weibull.R
\name{Mexcess_weibull}
\alias{Mexcess_weibull}
\title{Mean-Excess loss (excès-moyen) d'une loi Weibull}
\usage{
Mexcess_weibull(d, tau, beta)
}
\arguments{
\item{d}{déductible}

\item{tau}{tau}

\item{beta}{beta}
}
\description{
Mean-Excess loss (excès-moyen) d'une loi Weibull
}
