% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGF_PGF_binom.R
\name{MGF_binom}
\alias{MGF_binom}
\title{Moment and Probability Generating Functions of the Binomial distribution}
\usage{
MGF_binom(t, size, prob)
}
\arguments{
\item{t}{t}

\item{size}{Number of trials (0 or more).}

\item{prob}{Probability of success on each trial.}
}
\description{
Moment Generating Function (MGF) and Probability Generating
   Function (PGF) of the Binomial distribution with size \eqn{n}{n} and
   probability of success \eqn{p}{p}.
}
\details{
The Binomial distribution
}
\examples{

MGF_binom(t = 1, size = 3, prob = 0.5)
PGF_binom(t = 1, size = 3, prob = 0.5)

}
\seealso{
Other Binomial Distribution: 
\code{\link{E_binom}()},
\code{\link{Etronq_binom}()},
\code{\link{TVaR_binom}()},
\code{\link{V_binom}()},
\code{\link{VaR_binom}()}
}
\concept{Binomial Distribution}
