% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGF_exp.R
\name{MGF_exp}
\alias{MGF_exp}
\title{Moment Generating Function of the Exponential distribution}
\usage{
MGF_exp(t, rate = 1/scale, scale = 1/rate)
}
\arguments{
\item{t}{t.}

\item{rate}{\eqn{\beta}{beta} is the rate parameter, must be positive.}

\item{scale}{alternative parameterization to rate parameter, scale = 1 / rate.}
}
\description{
Moment Generating Function (MGF) of the Exponential
 distribution with rate parameter \eqn{\beta}{beta}.
}
\details{
The Exponential distribution with rate parameter \eqn{\beta}{b} has %
  density: \deqn{f\left(x\right) = \frac{1}{\beta}%
  \textrm{e}^{-\beta x}}{f(x) = b^a e^{-b x}} for %
  \eqn{x \in \mathcal{R}^+}{x > 0}, \eqn{\beta > 0}{b > 0}.
}
\examples{

MGF_exp(t = 1, rate = 5)

}
\seealso{
Other Exponential Distribution: 
\code{\link{E_exp}()},
\code{\link{Elim_exp}()},
\code{\link{Etronq_exp}()},
\code{\link{Mexcess_exp}()},
\code{\link{SL_exp}()},
\code{\link{TVaR_exp}()},
\code{\link{V_exp}()},
\code{\link{VaR_exp}()},
\code{\link{kthmoment_exp}()}
}
\concept{Exponential Distribution}
