% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGF_IG.R
\name{MGF_IG}
\alias{MGF_IG}
\title{Moment Generating Function (MGF) of the Inverse Gaussian distribution}
\usage{
MGF_IG(t, mu, beta = dispersion * mu^2, dispersion = beta/mu^2)
}
\arguments{
\item{t}{t}

\item{mu}{mu}

\item{beta}{beta = dispersion * mu^2}

\item{dispersion}{dispersion = beta / mu^2}
}
\description{
Moment Generating Function (MGF) of the Inverse Gaussian distribution
}
