% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGF_logarithmic.R
\name{MGF_logarithmic}
\alias{MGF_logarithmic}
\title{Moment Generating Function of the Logarithmic distribution}
\usage{
MGF_logarithmic(t, prob)
}
\arguments{
\item{t}{t.}

\item{prob}{probability parameter \eqn{\gamma}{gamma}.}
}
\description{
Moment Generating Function (MGF) of the Logarithmic
 distribution with probability parameter \eqn{\gamma}{gamma}.
}
\details{

}
\examples{

MGF_logarithmic(t = .2, prob = 0.50)

}
\seealso{
Other Logarithmic Distribution: 
\code{\link{E_logarithmique}()},
\code{\link{PGF_logarithmic}()},
\code{\link{V_logarithmique}()}
}
\concept{Logarithmic Distribution}
