% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGF_pois.R
\name{MGF_pois}
\alias{MGF_pois}
\title{Moment Generating Function of the Poisson distribution}
\usage{
MGF_pois(t, lambda)
}
\arguments{
\item{t}{t}

\item{lambda}{Rate parameter \eqn{\lambda}{lambda}.}
}
\description{
Moment Generating Function (PGF) of the Poisson
   distribution with rate parameter \eqn{\lambda}{lambda}.
}
\details{
The Poisson distribution with rate parameter \eqn{\lambda}{lam}
  has probability mass function :
  \deqn{Pr(X = k) = \frac{\lambda^k \textrm{e}^{-\lambda}}{k!}}{Pr(X = k) = (lam^k e^(-lam)) / k!}
  for \eqn{k = 0, 1, 2, \dots}{k = 0, 1, 2, ...}, and \eqn{\lambda > 0}{lam > 0}
}
\examples{

MGF_pois(t = 1, lambda = 3)

}
\seealso{
Other Poisson Distribution: 
\code{\link{Etronq_pois}()},
\code{\link{PGF_pois}()},
\code{\link{TVaR_pois}()}
}
\concept{Poisson Distribution}
