% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentsBox.R
\name{momentsBox}
\alias{momentsBox}
\title{Interactive moments visualization (server side)}
\usage{
momentsBox(input, output, session, law, lang)
}
\arguments{
\item{input}{input for server side.}

\item{output}{output for server side.}

\item{session}{session for server side.}

\item{law}{Distribution to visualize, one of ...}

\item{lang}{Internal function to ensure translation works and input is communicated between modules.}
}
\value{
Server function for the moments module.
 Should not be run directly.
}
\description{
Interactive moments visualization (server side)
}
\keyword{internal}
