% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentsBox.R
\name{momentsBoxUI}
\alias{momentsBoxUI}
\title{Interactive moments visualization (UI side)}
\usage{
momentsBoxUI(id)
}
\arguments{
\item{id}{id of module}
}
\value{
UI function for the moments module.
 Should not be run directly.
}
\description{
Interactive moments visualization (UI side)
}
\keyword{internal}
