% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_BNCOMP.R
\name{p_BNCOMP}
\alias{p_BNCOMP}
\title{Répartition d'une loi binomiale négative composée}
\usage{
p_BNCOMP(x, r, q, shape, rate, ko, distr_severity = "Gamma")
}
\arguments{
\item{x}{x}

\item{r}{r pour la binomiale négative (nombre d'échecs).}

\item{q}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{ko}{borne pour la sommation de la fonction de répartition}

\item{distr_severity}{choix de distribution de sévérité. Gamma ou Lognormale}
}
\description{
Répartition d'une loi binomiale négative composée
}
\details{
Cette formule utilise aussi un choix de 2 distributions pour la sévérité; soit la Gamma (par défaut) ou la Lognormale.
}
