% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_negbinom.R
\name{p_negbinom}
\alias{p_negbinom}
\title{Fonction de répartition de la loi binomiale négative}
\usage{
p_negbinom(k, r, p = (1/(1 + beta)), beta = ((1 - p)/p),
  nb_tries = F, lower.tail = F)
}
\arguments{
\item{k}{Nombre de échecs (ou succès) jusqu'où sommer.}

\item{r}{Nombre de succès visé.}

\item{p}{Probabilité de succès.}

\item{beta}{Deuxième paramétrisation de la loi binomiale négative tel que beta = (1 - p) / p.}

\item{nb_tries}{Stipule si la binomiale est une du nombre d'essais pour le r-ème succès, ou du nombre d'échecs pour le r-ème succès (par défaut).}

\item{lower.tail}{Si vrai (défaut) alors fonction de répartition. Si faux, de survie.}
}
\description{
Fonction de répartition de la loi binomiale négative
}
