% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_unifD.R
\name{p_unifD}
\alias{p_unifD}
\title{Cumulative probability mass function of the (discrete) Uniform distribution}
\usage{
p_unifD(q, min = 0, max = 1)
}
\arguments{
\item{q}{quantile.}

\item{min, max}{lower and upper limits of the distribution. Must be finite.}
}
\description{
Cumulative probability mass function of the (discrete) Uniform
 distribution with min \eqn{a} and max \eqn{b}.
}
\details{
The (discrete) uniform distribution with min and max parameters \eqn{a}
  and \eqn{b} respectively has density:
  \deqn{\textrm{Pr}\left(X = x \right) = \frac{1}{b - a + 1}}{Pr(X = x) = 1 / (b - a + 1)}
  for \eqn{x \in \{a, a + 1, \dots, b - 1, b\}}{x = a, a + 1, ..., b - 1, b}.
}
\examples{

# With scale parameter
p_unifD(q = 2, min = 2, max = 5)

}
\concept{Discrete Uniform Distribution}
