% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersBox.R
\name{parametersBox}
\alias{parametersBox}
\title{Interactive parameter chooser (server side)}
\usage{
parametersBox(input, output, session, law, lang)
}
\arguments{
\item{input}{input for server side.}

\item{output}{output for server side.}

\item{session}{session for server side.}

\item{law}{Distribution to visualize, one of ...}

\item{lang}{Internal function to ensure translation works and input is communicated between modules.}
}
\value{
Server function for the parameter module.
 Should not be run directly.
}
\description{
Interactive parameter chooser (server side)
}
\keyword{internal}
