% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersBoxCopulas.R
\name{parametersBoxCopulasUI}
\alias{parametersBoxCopulasUI}
\title{Interactive parameter chooser (UI side) for copulas}
\usage{
parametersBoxCopulasUI(id)
}
\arguments{
\item{id}{id of module}
}
\value{
UI function for the parameter module of the Copula Shiny
 function. Should not be run directly.
}
\description{
Interactive parameter chooser (UI side) for copulas
}
\keyword{internal}
