% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersBox.R
\name{parametersBoxUI}
\alias{parametersBoxUI}
\title{Interactive parameter chooser (UI side)}
\usage{
parametersBoxUI(id)
}
\arguments{
\item{id}{id of module}
}
\value{
UI function for the parameter module.
 Should not be run directly.
}
\description{
Interactive parameter chooser (UI side)
}
\keyword{internal}
