% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlang.R
\name{perlang}
\alias{perlang}
\title{Fonction de répartition de la loi Erlang}
\usage{
perlang(x, n, b, lower.tail = T)
}
\arguments{
\item{n}{n}

\item{b}{beta}

\item{k}{niveau de confiance désiré}
}
\description{
Fonction de répartition de la loi Erlang
}
