% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGF_binom.R
\name{PGF_binom}
\alias{PGF_binom}
\title{Probability Generating Function of the Binomial distribution}
\usage{
PGF_binom(t, size, prob)
}
\arguments{
\item{t}{t}

\item{size}{Number of trials (0 or more).}

\item{prob}{Probability of success on each trial.}
}
\description{
Probability Generating Function (PGF) of the Binomial
 distribution with size \eqn{n}{n} and probability of success \eqn{p}{p}.
}
\details{
The Binomial distribution
}
\examples{

PGF_binom(t = 1, size = 3, prob = 0.5)

}
\seealso{
Other Binomial Distribution: 
\code{\link{E_binom}()},
\code{\link{Etronq_binom}()},
\code{\link{MGF_binom}()},
\code{\link{TVaR_binom}()},
\code{\link{V_binom}()},
\code{\link{VaR_binom}()}
}
\concept{Binomial Distribution}
