% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGF_pois.R
\name{PGF_pois}
\alias{PGF_pois}
\title{Probability Generating Function of the Poisson distribution}
\usage{
PGF_pois(t, lambda)
}
\arguments{
\item{t}{t}

\item{lambda}{Rate parameter \eqn{\lambda}{lambda}.}
}
\description{
Probability Generating Function (PGF) of the Poisson
   distribution with rate parameter \eqn{\lambda}{lambda}.
}
\details{

}
\examples{

PGF_pois(t = 1, lambda = 3)

}
\seealso{
Other Poisson Distribution: 
\code{\link{Etronq_pois}()},
\code{\link{MGF_pois}()},
\code{\link{TVaR_pois}()}
}
\concept{Poisson Distribution}
