% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pIG.R
\name{pIG}
\alias{pIG}
\title{Envelope de la densité de la loi inverse gaussienne}
\usage{
pIG(q, mu, beta = dispersion * mu^2, dispersion = beta/mu^2,
  lower.tail = T)
}
\arguments{
\item{q}{quantile}

\item{mu}{mu}

\item{beta}{beta = dispersion * mu^2}

\item{dispersion}{dispersion = beta / mu^2}
}
\description{
Envelope de la densité de la loi inverse gaussienne
}
\details{
Wrapper pour avoir une notation cohérente avec le reste des fonctions IG, mais utilise la fonction de densité du package ACTUAR.
}
