% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pllogis.R
\name{pllogis}
\alias{pllogis}
\title{Cumulative density function of the Loglogistic distribution}
\usage{
pllogis(q, shape, rate = 1/scale, scale = 1/rate, lower.tail = TRUE)
}
\arguments{
\item{q}{quantile.}

\item{shape}{shape parameter \eqn{\tau}{tau}, must be positive integer.}

\item{rate}{alternative parameterization the scale parameter, rate = 1 / scale.}

\item{scale}{\eqn{\lambda}{lambda} rate parameter, must be positive.}

\item{lower.tail}{logical; if \\code{TRUE} (default) probabilities are Pr(M <= k), otherwise, Pr(M > k).}
}
\value{
Function :
 \itemize{
 \item{\code{\link{dllogis}}}{ gives the density function.}
 \item{\code{\link{pllogis}}}{ gives the cumulative density function.}
 \item{\code{\link{E_llogis}}}{ gives the expected value.}
 \item{\code{\link{V_llogis}}}{ gives the variance.}
 \item{\code{\link{kthmoment_llogis}}}{ gives the kth moment.}
 \item{\code{\link{Etrunc_llogis}}}{ gives the truncated mean.}
 \item{\code{\link{SL_llogis}}}{ gives the stop-loss.}
 \item{\code{\link{Elim_llogis}}}{ gives the limited mean.}
 \item{\code{\link{Mexcess_llogis}}}{ gives the mean excess loss.}
 \item{\code{\link{TVaR_llogis}}}{ gives the Tail Value-at-Risk.}
 \item{\code{\link{VaR_llogis}}}{ gives the Value-at-Risk.}
 }
 Invalid parameter values will return an error detailing which parameter is problematic.
}
\description{
Cumulative density function of the Loglogistic distribution
 with shape parameter \eqn{\tau}{tau} and scale parameter \eqn{\lambda}{lambda}.
}
\details{
The Loglogistic distribution with shape parameter \eqn{\tau}{t} and scale parameter
  \eqn{\lambda}{lam} has density:
  \deqn{\frac{\tau \lambda^\tau x^{\tau -1}}{(\lambda^{\tau }+x^{\tau })^{2}}}{f(x) = t lam^t x^(t - 1) / (lam^t + x^t)^2}
  for \eqn{x \in \mathcal{R}^+}{x > 0}, \eqn{\lambda, \tau > 0}{lam, t > 0}.
}
\examples{

# With scale parameter
pllogis(q = 3, shape = 3, scale = 5)

# With rate parameter
pllogis(q = 3, shape = 3, rate = 0.2)

# Survival function
pllogis(q = 3, shape = 3, rate = 0.2, lower.tail = FALSE)

}
\seealso{
Other Loglogistic Distribution: 
\code{\link{E_llogis}()},
\code{\link{Elim_llogis}()},
\code{\link{Etrunc_llogis}()},
\code{\link{Mexcess_llogis}()},
\code{\link{SL_llogis}()},
\code{\link{TVaR_llogis}()},
\code{\link{V_llogis}()},
\code{\link{VaR_llogis}()},
\code{\link{dllogis}()},
\code{\link{kthmoment_llogis}()}
}
\concept{Loglogistic Distribution}
