% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punifD.R
\name{punifD}
\alias{punifD}
\title{Cumulative probability mass function of the (discrete) Uniform Distribution}
\usage{
punifD(q, min, max)
}
\arguments{
\item{q}{quantile.}

\item{min, max}{lower and upper limits of the distribution. Must be finite.#'}
}
\description{
Cumulative probability mass function of the (discrete) Uniform distribution
 with min \eqn{a} and max \eqn{b}.
}
\details{
The (discrete) uniform distribution with min and max parameters \eqn{a}
  and \eqn{b} respectively has density:
  \deqn{\textrm{Pr}\left(X = x \right) = \frac{1}{b - a + 1}}{Pr(X = x) = 1 / (b - a + 1)}
  for \eqn{x \in \{a, a + 1, \dots, b - 1, b\}}{x = a, a + 1, ..., b - 1, b}.
}
\examples{

# With scale parameter
punifD(q = 2, min = 2, max = 5)

}
\seealso{
Other Discrete Uniform Distribution: \code{\link{E_unifD}},
  \code{\link{V_unifD}}, \code{\link{dunifD}}
}
\concept{Discrete Uniform Distribution}
