% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgompertz.R
\name{qgompertz}
\alias{qgompertz}
\title{fonction quantile de la loi Gompertz}
\usage{
qgompertz(q, x = 0, beta, gam, Tx = T)
}
\arguments{
\item{q}{quantile (va de 0 à 1)}

\item{x}{âge ateint}

\item{beta}{paramètre beta}

\item{gam}{paramètre gamma}

\item{Tx}{Paramètre Tx, si on désire avoir la distribution de Tx.}
}
\description{
fonction quantile de la loi Gompertz
}
