% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskMeasuresBox.R
\name{riskMeasuresBoxUI}
\alias{riskMeasuresBoxUI}
\title{Interactive risk measures visualization (UI side)}
\usage{
riskMeasuresBoxUI(id)
}
\arguments{
\item{id}{id of module}
}
\value{
UI function for the risk measures module.
 Should not be run directly.
}
\description{
Interactive risk measures visualization (UI side)
}
\keyword{internal}
