% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgompertz.R
\name{sgompertz}
\alias{sgompertz}
\title{fonction de survie de la loi Gompertz}
\usage{
sgompertz(t, x = 0, beta, gam, Tx = T)
}
\arguments{
\item{t}{durée de vie T_x}

\item{x}{âge ateint}

\item{beta}{paramètre beta}

\item{gam}{paramètre gamma}

\item{Tx}{Paramètre Tx, si on désire avoir la distribution de Tx.}
}
\description{
fonction de survie de la loi Gompertz
}
