% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedCopulaHexPlot.R
\name{simulatedCopulaHexPlot}
\alias{simulatedCopulaHexPlot}
\title{Plot of simulated observations for copula}
\usage{
simulatedCopulaHexPlot(
  copula,
  numberSimulations = 10000,
  seed = 42,
  dependencyParameter
)
}
\arguments{
\item{copula}{Copula to visualize, presently one of :
\itemize{
 \item{"Independant": }{ Independant copula.}
 \item{"FrechetLowerBound": }{ Fréchet Lower Bound copula.}
 \item{"FrechetUpperBound": }{ Fréchet Upper Bound copula.}
 \item{"BivariateEFGM": }{ Eyraud-Farlie-Gumbel-Morgenstern copula.}
 \item{"Frechet": }{ Fréchet copula.}
 \item{"BivariateCA": }{ Bivariate Cuadras-Augé copula.}
 \item{"BivariateMO": }{ Bivariate Marhsall-Olkin copula.}
 \item{"BivariateAMH": }{ Bivariate Ali-Mikhail-Haq copula.}
 \item{"BivariateClayton": }{ Bivariate Clayton copula.}
 \item{"BivariateFrank": }{ Bivariate Frank copula.}
 \item{"BivariateGumbel": }{ Bivariate Gumbel copula.}
}}

\item{numberSimulations}{Number of simulations.}

\item{seed}{Simulation seed, 42 by default.}

\item{dependencyParameter}{correlation parameter.}
}
\value{
Returns plot of simulated observations for the chosen copula.
}
\description{
Plot of simulated observations for copula
}
