% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_burr.R
\name{SL_burr}
\alias{SL_burr}
\title{Stop-loss d'une loi Burr}
\usage{
SL_burr(d, alpha, lam, tau)
}
\arguments{
\item{d}{montant d de la franchise}

\item{alpha}{alpha}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Stop-loss d'une loi Burr
}
