% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_erlang.R
\name{SL_erlang}
\alias{SL_erlang}
\title{Stop-loss d'une loi exponentielle}
\usage{
SL_erlang(d, n, b)
}
\arguments{
\item{d}{montant d de la franchise}

\item{n}{n}

\item{b}{beta}
}
\description{
Stop-loss d'une loi exponentielle
}
