% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_gamma.R, R/TVaR_unif.R
\name{SL_gamma}
\alias{SL_gamma}
\alias{SL_gamma}
\title{Stop-loss d'une loi Gamma}
\usage{
SL_gamma(d, a, b)

SL_gamma(d, a, b)
}
\arguments{
\item{d}{déductible}

\item{a}{alpha}

\item{b}{beta}

\item{d}{déductible}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Stop-loss d'une loi Gamma

Stop-loss d'une loi gamma
}
