% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_IG.R
\name{SL_IG}
\alias{SL_IG}
\title{Stop-loss de la loi inverse gaussienne}
\usage{
SL_IG(d, mu, beta = dispersion * mu^2, dispersion = beta/mu^2)
}
\arguments{
\item{d}{montant d de la franchise}

\item{mu}{mu}

\item{beta}{beta = dispersion * mu^2}

\item{dispersion}{dispersion = beta / mu^2}
}
\description{
Stop-loss de la loi inverse gaussienne
}
