% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_llogis.R
\name{SL_llogis}
\alias{SL_llogis}
\title{Stop-loss d'une loi log-logistique}
\usage{
SL_llogis(d, lam, tau)
}
\arguments{
\item{d}{montant d de la franchise}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Stop-loss d'une loi log-logistique
}
