% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_lnorm.R
\name{SL_lnorm}
\alias{SL_lnorm}
\title{Stop-Loss d'une loi lognormale}
\usage{
SL_lnorm(d, mu, sig)
}
\arguments{
\item{d}{déductible}

\item{mu}{mu}

\item{sig}{sigma}
}
\description{
Stop-Loss d'une loi lognormale
}
