% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_pareto.R
\name{SL_pareto}
\alias{SL_pareto}
\title{Stop-Loss d'une loi Pareto}
\usage{
SL_pareto(d, alph, lam)
}
\arguments{
\item{d}{déductible}

\item{alph}{alpha}

\item{lam}{lambda}
}
\description{
Stop-Loss d'une loi Pareto
}
