% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_weibull.R
\name{SL_weibull}
\alias{SL_weibull}
\title{Stop-loss d'une loi Weibull}
\usage{
SL_weibull(d, tau, beta)
}
\arguments{
\item{d}{montant d de la franchise}

\item{tau}{tau}

\item{beta}{beta}
}
\description{
Stop-loss d'une loi Weibull
}
