% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_beta.R
\name{TVaR_beta}
\alias{TVaR_beta}
\title{Tail Value-at-Risk  d'une loi Beta}
\usage{
TVaR_beta(k, a, b)
}
\arguments{
\item{k}{kappa (doit être entre 0 et 1)}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Tail Value-at-Risk  d'une loi Beta
}
