% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_BINCOMP.R
\name{TVaR_BINCOMP}
\alias{TVaR_BINCOMP}
\title{Tail Value-at-risk d'une loi binomiale composée}
\usage{
TVaR_BINCOMP(kappa, n, q, shape, rate, vark, ko,
  distr_severity = "Gamma")
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{n}{n pour la binomiale}

\item{q}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{vark}{Value at Risk (VaR) calculé au même niveau de confiance kappa}

\item{ko}{borne pour la somme dans le calcul de la TVaR}

\item{distr_severity}{choix de distribution de sévérité. Gamma ou Lognormale}
}
\description{
Tail Value-at-risk d'une loi binomiale composée
}
\details{
Cette formule nécessite la formule de la VaR_bincomp (déjà installée avec le package tvarPackage). Cette formule utilise aussi un choix de (présentement) 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
