% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_binom.R
\name{TVaR_binom}
\alias{TVaR_binom}
\title{Tail Value-at-risk de la loi binomiale}
\usage{
TVaR_binom(kappa, n, p)
}
\arguments{
\item{kappa}{niveau de confiance désiré}

\item{n}{nombres d'essais de Bernouilli}

\item{p}{probabilité de succès (pour chaque expérience indépendantes)}
}
\description{
Tail Value-at-risk de la loi binomiale
}
