% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_BNCOMP.R
\name{TVaR_BNCOMP}
\alias{TVaR_BNCOMP}
\title{Tail Value-at-risk d'une loi binomiale négative composée}
\usage{
TVaR_BNCOMP(kap, vark, size, prob, shape, rate, distr_severity = "Gamma", ko)
}
\arguments{
\item{kap}{Niveau de confiance désiré}

\item{vark}{Value at Risk (VaR) calculé au même niveau de confiance kap}

\item{size}{r pour la binomiale négative}

\item{prob}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{distr_severity}{choice of severity distribution Gamma (default)ss}

\item{ko}{borne pour la somme dans le calcul de la TVaR.}
}
\description{
Tail Value-at-risk d'une loi binomiale négative composée
}
\details{
Cette formule nécessite la formule de la VaR_bncomp (déjà installée avec le package tvarPackage). Cette formule utilise aussi un choix de (présentement) 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
