% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_burr.R
\name{TVaR_burr}
\alias{TVaR_burr}
\title{Tail Value-at-risk d'une loi Burr}
\usage{
TVaR_burr(kappa, var, alpha, lam, tau)
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{var}{Value at Risk (VaR) calculé au même niveau de risque}

\item{alpha}{alpha}

\item{lam}{lam}

\item{tau}{tau}
}
\description{
Tail Value-at-risk d'une loi Burr
}
\details{
Cette formule nécessite la formule de la VaR_burr (déjà installée avec le package tvarPackage)
}
