% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_erlang.R
\name{TVaR_erlang}
\alias{TVaR_erlang}
\title{Tail Value-at-risk d'une loi Erlang}
\usage{
TVaR_erlang(vark, k, n, b)
}
\arguments{
\item{vark}{Value at Risk (VaR) calculé au même niveau de confiance k}

\item{k}{Niveau de confiance désiré}

\item{alpha}{alpha}

\item{lam}{lam}

\item{tau}{tau}
}
\description{
Tail Value-at-risk d'une loi Erlang
}
\details{
Cette formule nécessite la formule de la VaR_erlang (déjà installée avec le package tvarPackage)
}
