% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_exp.R
\name{TVaR_exp}
\alias{TVaR_exp}
\title{Tail Value-at-risk d'une loi exponentielle}
\usage{
TVaR_exp(kap, b)
}
\arguments{
\item{kap}{pourcentage de confiance désiré}

\item{b}{beta}
}
\description{
Tail Value-at-risk d'une loi exponentielle
}
