% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_gamma.R
\name{TVaR_gamma}
\alias{TVaR_gamma}
\title{Tail Value-at-risk d'une loi gamma}
\usage{
TVaR_gamma(kappa, a, b)
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Tail Value-at-risk d'une loi gamma
}
\details{
Cette formule nécessite la formule VaR_gamma (déjà installée avec le package tvarPackage)
}
