% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_IG.R
\name{TVaR_IG}
\alias{TVaR_IG}
\title{Tail Value-at-risk d'une loi inverse gaussienne}
\usage{
TVaR_IG(kap, mu, beta = dispersion * mu^2, dispersion = beta/mu^2, vark)
}
\arguments{
\item{kap}{Niveau de confiance désiré}

\item{mu}{mu}

\item{beta}{beta, = dispersion * mu^2}

\item{dispersion}{dispersion, = beta / mu^2}

\item{vark}{Value at Risk (VaR) calculé au même niveau de confiance kap}
}
\description{
Tail Value-at-risk d'une loi inverse gaussienne
}
\details{
Cette formule nécessite la formule de la VaR_IG (déjà installée avec le package tvarPackage)
}
