% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_lnorm.R
\name{TVaR_lnorm}
\alias{TVaR_lnorm}
\title{Tail Value-at-Risk d'une loi lognormale}
\usage{
TVaR_lnorm(kappa, mu, sig)
}
\arguments{
\item{kappa}{pourcentage de confiance désiré}

\item{mu}{mu}

\item{sig}{sigma}
}
\description{
Tail Value-at-Risk d'une loi lognormale
}
