% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_pareto.R
\name{TVaR_pareto}
\alias{TVaR_pareto}
\title{Tail Value-at-risk d'une loi Pareto}
\usage{
TVaR_pareto(kap, alph, lam)
}
\arguments{
\item{kap}{kap}

\item{alph}{alpha}

\item{lam}{lambda}
}
\description{
Tail Value-at-risk d'une loi Pareto
}
