% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_weibull.R
\name{TVaR_weibull}
\alias{TVaR_weibull}
\title{Tail Value-at-risk d'une loi Weibull}
\usage{
TVaR_weibull(kappa, tau, beta)
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{tau}{tau}

\item{beta}{beta}
}
\description{
Tail Value-at-risk d'une loi Weibull
}
