% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaRtx_gompertz.R
\name{TVaRtx_gompertz}
\alias{TVaRtx_gompertz}
\title{Tail Value-at-risk de Tx pour la loi Gompertz}
\usage{
TVaRtx_gompertz(kappa, x = 0, beta, gam)
}
\arguments{
\item{kappa}{niveau de confiance pour l'évaluation de la TVaR}

\item{x}{l'âge atteint de l'assuré.}

\item{beta}{paramètre beta}

\item{gam}{paramètre gamma}
}
\description{
Tail Value-at-risk de Tx pour la loi Gompertz
}
\details{
Cette fonction nécessite d'avoir loadé le package `tvarPackage` au complet, car elle fait appel à une autre fonction (celle équivalente à la VaR) `qgompertz`.
}
