% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_BINCOMP.R
\name{V_BINCOMP}
\alias{V_BINCOMP}
\title{Variance d'une loi binomiale composée}
\usage{
V_BINCOMP(n, q, shape, rate, distr_severity = "Gamma")
}
\arguments{
\item{n}{n pour la binomiale}

\item{q}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{distr_severity}{choix de distribution de sévérité. Gamma (par défaut) ou Lognormale}
}
\description{
Variance d'une loi binomiale composée
}
\details{
Cette formule utilise aussi un choix de 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
