% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_binom.R
\name{V_binom}
\alias{V_binom}
\title{Variance of the Binomial distribution}
\usage{
V_binom(size, prob)
}
\arguments{
\item{size}{Number of trials (0 or more).}

\item{prob}{Probability of success on each trial.}
}
\description{
Variance of the Binomial distribution with size
 \eqn{n}{n} and probability of success \eqn{p}{p}.
}
\details{
The Binomial distribution
}
\examples{

V_binom(size = 3, prob = 0.5)

}
\seealso{
Other Binomial Distribution: 
\code{\link{E_binom}()},
\code{\link{Etronq_binom}()},
\code{\link{MGF_binom}()},
\code{\link{PGF_binom}()},
\code{\link{TVaR_binom}()},
\code{\link{VaR_binom}()}
}
\concept{Binomial Distribution}
