% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_hyper.R
\name{V_hyper}
\alias{V_hyper}
\title{Variance d'une loi hypergéométrique}
\usage{
V_hyper(N, m, n)
}
\arguments{
\item{N}{nombre total}

\item{m}{nombre de type 1}

\item{n}{taille de l'échantillon}
}
\description{
Variance d'une loi hypergéométrique
}
