% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_IG.R
\name{V_IG}
\alias{V_IG}
\title{Variance of the Inverse Gaussian distribution}
\usage{
V_IG(mean, shape = dispersion * mean^2, dispersion = shape/mean^2)
}
\arguments{
\item{mean}{mean (location) parameter \eqn{\mu}{mu}, must be positive.}

\item{shape}{shape parameter \eqn{\beta}{beta}, must be positive.}

\item{dispersion}{alternative parameterization to the shape parameter, dispersion = 1 / rate.}
}
\description{
Variance of the Inverse Gaussian distribution with
 mean \eqn{\mu}{mu} and shape parameter \eqn{\beta}{beta}.
}
\details{
The Pareto distribution with rate parameter \eqn{\lambda}{lam} as well as shape
  parameter \eqn{\alpha}{a} has density: \deqn{f\left(x\right) = \frac{\alpha%
  \lambda^{\alpha}}{(\lambda + x)^{\alpha + 1}}}{f(x) = (a lam^a)/ (lam + x)^(a + 1)}
  for \eqn{x \in \mathcal{R}^+}{x > 0}, \eqn{\alpha, \lambda > 0}{a, lam > 0}.
}
\examples{

V_IG(mean = 2, shape = 5)

}
\seealso{
Other Inverse Gaussian Distribution: 
\code{\link{E_IG}()},
\code{\link{Elim_IG}()},
\code{\link{Etronq_IG}()},
\code{\link{MGF_IG}()},
\code{\link{Mexcess_IG}()},
\code{\link{SL_IG}()},
\code{\link{TVaR_IG}()},
\code{\link{VaR_IG}()}
}
\concept{Inverse Gaussian Distribution}
