% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_nbinom.R
\name{V_nbinom}
\alias{V_nbinom}
\title{Variance d'une loi binomiale négative
Calcul la variance d'une binomiale négative}
\usage{
V_nbinom(r, q)
}
\arguments{
\item{r}{nombre d'essais}

\item{q}{probabilité d'un succès indépendant}
}
\description{
Variance d'une loi binomiale négative
Calcul la variance d'une binomiale négative
}
