% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_unifD.R
\name{V_unifD}
\alias{V_unifD}
\title{Variance of the (discrete) Uniform distribution}
\usage{
V_unifD(min = 0, max = 1)
}
\arguments{
\item{min, max}{lower and upper limits of the distribution. Must be finite.}
}
\description{
Variance of the (discrete) Uniform distribution
 with min \eqn{a} and max \eqn{b}.
}
\examples{

# With scale parameter
V_unifD(min = 2, max = 5)

}
