% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_beta.R
\name{VaR_beta}
\alias{VaR_beta}
\title{Value-at-Risk of the Beta distribution}
\usage{
VaR_beta(kap, shape1, shape2)
}
\arguments{
\item{kap}{kap (doit être entre 0 et 1)}

\item{shape1}{alpha}

\item{shape2}{beta}
}
\description{
Value-at-Risk of the Beta distribution
}
