% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_BINCOMP.R
\name{VaR_BINCOMP}
\alias{VaR_BINCOMP}
\title{Value-at-Risk of the compound Binomiale distribution}
\usage{
VaR_BINCOMP(kap, size, prob, shape, rate, distr_severity = "Gamma", ko)
}
\arguments{
\item{kap}{Niveau de confiance désiré}

\item{size}{n pour la binomiale}

\item{prob}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{distr_severity}{choix de distribution de sévérité. Gamma uniquement pour l'instant.}

\item{ko}{borne pour l'optimisation numérique de la VaR}
}
\description{
Value-at-Risk of the compound Binomiale distribution
}
