% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_binom.R
\name{VaR_binom}
\alias{VaR_binom}
\title{Value-at-risk de la loi binomiale}
\usage{
VaR_binom(kappa, n, p)
}
\arguments{
\item{kappa}{niveau de confiance désiré}

\item{n}{nombres d'essais de Bernouilli}

\item{p}{probabilité de succès (pour chaque expérience indépendantes)}
}
\description{
Value-at-risk de la loi binomiale
}
