% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_BNCOMP.R
\name{VaR_BNCOMP}
\alias{VaR_BNCOMP}
\title{Value-at-risk d'une loi binomiale négative composée}
\usage{
VaR_BNCOMP(kap, r, q, shape, rate, ko = 300, distr_severity = "Gamma")
}
\arguments{
\item{kap}{Niveau de confiance désiré}

\item{r}{r pour la binomiale négative (nombre d'échecs).}

\item{q}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{ko}{borne pour l'optimisation numérique de la VaR}

\item{distr_severity}{choix de distribution de sévérité.}
}
\description{
Value-at-risk d'une loi binomiale négative composée
}
