% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_exp.R
\name{VaR_exp}
\alias{VaR_exp}
\title{Value-at-risk d'une loi exponentielle}
\usage{
VaR_exp(kappa, b)
}
\arguments{
\item{kappa}{pourcentage de confiance désiré}

\item{b}{beta}
}
\description{
Value-at-risk d'une loi exponentielle
}
\details{
La VaR est la fonction quantile de la v.a (ici, de la loi exponentielle)
}
