% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_FGEN.R
\name{VaR_FGEN}
\alias{VaR_FGEN}
\title{Value-at-Risk of the F-Generalized distribution}
\usage{
VaR_FGEN(x, kap, lambda, alpha, tau)
}
\arguments{
\item{x}{x}

\item{kap}{niveau de confiance désiré}

\item{lambda}{lambda, > 0}

\item{alpha}{alpha, >0}

\item{tau}{tau, >0}
}
\description{
Value-at-Risk of the F-Generalized distribution
}
