% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_IG.R
\name{VaR_IG}
\alias{VaR_IG}
\title{Value-at-risk de la loi inverse gaussienne}
\usage{
VaR_IG(k, mu, beta = dispersion * mu^2, dispersion = beta/mu^2)
}
\arguments{
\item{k}{niveau de confiance désiré}

\item{mu}{mu}

\item{beta}{beta = dispersion * mu^2}

\item{dispersion}{dispersion = beta / mu^2}
}
\description{
Value-at-risk de la loi inverse gaussienne
}
\details{
Wrapper pour avoir une notation cohérente avec le reste des fonctions IG, mais utilise la fonction de densité du package ACTUAR.
}
