% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_llogis.R
\name{VaR_llogis}
\alias{VaR_llogis}
\title{Value-at-risk d'une loi log-logstique}
\usage{
VaR_llogis(kappa, lam, tau)
}
\arguments{
\item{kappa}{niveau de confiance désiré}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Value-at-risk d'une loi log-logstique
}
