% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_lnorm.R
\name{VaR_lnorm}
\alias{VaR_lnorm}
\title{Value-at-Risk d'une loi lognormale}
\usage{
VaR_lnorm(kappa, mu, sig)
}
\arguments{
\item{kappa}{pourcentage de confiance désirée}

\item{mu}{mu}

\item{sig}{sigma}
}
\description{
Value-at-Risk d'une loi lognormale
}
