% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_norm.R
\name{VaR_norm}
\alias{VaR_norm}
\title{Value-at-Risk d'une loi normale}
\usage{
VaR_norm(kappa, mu, sig)
}
\arguments{
\item{kappa}{pourcentage de confiance désiré}

\item{mu}{mu}

\item{sig}{sigma}
}
\description{
Value-at-Risk d'une loi normale
}
