% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_pareto.R
\name{VaR_pareto}
\alias{VaR_pareto}
\title{Value-at-risk d'une loi Pareto}
\usage{
VaR_pareto(kappa, alph, lam)
}
\arguments{
\item{kappa}{kappa}

\item{alph}{alpha}

\item{lam}{lambda}
}
\description{
Value-at-risk d'une loi Pareto
}
