% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_PCOMP.R
\name{VaR_PCOMP}
\alias{VaR_PCOMP}
\title{Fonction de répartition d'une loi poisson composée}
\usage{
VaR_PCOMP(kappa, lambda, shape, rate, ko = 300, distr_severity = "Gamma")
}
\arguments{
\item{kappa}{niveau de confiance désiré}

\item{lambda}{paramètre lambda pour la poisson}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{ko}{borne pour l'optimisation de la VaR}

\item{distr_severity}{choix de distribution de sévérité.}
}
\description{
Fonction de répartition d'une loi poisson composée
}
