% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_unif.R
\name{VaR_unif}
\alias{VaR_unif}
\title{Value-at-Risk of the Uniform distribution}
\usage{
VaR_unif(kap, min, max)
}
\arguments{
\item{kap}{probability.}

\item{max}{maximum}
}
\description{
Value-at-Risk of the Uniform distribution with shape
 parameters \eqn{\alpha}{alpha} and \eqn{\beta}{beta}.
}
