% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_weibull.R
\name{VaR_weibull}
\alias{VaR_weibull}
\title{Value-at-risk de la loi Weibull}
\usage{
VaR_weibull(k, tau, beta)
}
\arguments{
\item{k}{niveau de confiance désiré}

\item{tau}{tau}

\item{beta}{beta}
}
\description{
Value-at-risk de la loi Weibull
}
