% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vtx_gompertz.R
\name{Vtx_gompertz}
\alias{Vtx_gompertz}
\title{Variance de Tx pour la loi Gompertz}
\usage{
Vtx_gompertz(x = 0, beta, gam, floor = T)
}
\arguments{
\item{x}{âge ateint}

\item{beta}{paramètre beta}

\item{gam}{paramètre gamma}

\item{floor}{Pour valider quelles techniques on utilise.}
}
\description{
Variance de Tx pour la loi Gompertz
}
\details{
pour le paramètre `floor`, on peut utiliser `T` si on veut utiliser la partie entière. `F` fera usage de la fonction `integrate` de R.
}
