% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunifD.R
\name{dunifD}
\alias{dunifD}
\title{Densité d'une loi uniforme discrète}
\usage{
dunifD(x, a, b)
}
\arguments{
\item{x}{x n'a aucun impact pour une loi uniforme par définition}

\item{a}{borne inférieur}

\item{b}{borne supérieur}
}
\description{
Densité d'une loi uniforme discrète
}
\details{
loi uniforme discrète
}
