% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunifD.R
\name{dunifD}
\alias{dunifD}
\title{Probability mass function of the (discrete) Uniform distribution}
\usage{
dunifD(x, min = 0, max = 1)
}
\arguments{
\item{x}{quantile. By definition, it has no impact on the uniform distribution. Set to 1 by default.}

\item{min, max}{lower and upper limits of the distribution. Must be finite.}

\item{q}{quantile.}
}
\description{
Probability mass function of the (discrete) Uniform
 distribution with min \eqn{a} and max \eqn{b}.
}
\details{
The (discrete) uniform distribution with min and max parameters \eqn{a}
  and \eqn{b} respectively has density:
  \deqn{\textrm{Pr}\left(X = x \right) = \frac{1}{b - a + 1}}{Pr(X = x) = 1 / (b - a + 1)}
  for \eqn{x \in \{a, a + 1, \dots, b - 1, b\}}{x = a, a + 1, ..., b - 1, b}.
}
\examples{

# With scale parameter
dunifD(x = 2, min = 2, max = 5)

}
\seealso{
Other Discrete Uniform Distribution: 
\code{\link{E_unifD}()},
\code{\link{V_unifD}()},
\code{\link{punifD}()}
}
\concept{Discrete Uniform Distribution}
