% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{E_binom}
\alias{E_binom}
\alias{V_binom}
\alias{Etronq_binom}
\alias{VaR_binom}
\alias{TVaR_binom}
\alias{PGF_binom}
\alias{MGF_binom}
\title{Binomial Distribution}
\usage{
E_binom(size, prob)

V_binom(size, prob)

Etronq_binom(d, size, prob, less.than.d = TRUE)

VaR_binom(kap, size, prob)

TVaR_binom(kap, size, prob)

PGF_binom(t, size, prob)

MGF_binom(t, size, prob)
}
\arguments{
\item{size}{Number of trials (0 or more).}

\item{prob}{Probability of success in each trial.}

\item{d}{cut-off value.}

\item{less.than.d}{logical; if \\code{TRUE} (default) truncated mean for values <= d, otherwise, for values > d.}

\item{kap}{probability.}

\item{t}{t.}
}
\description{
Binomial distribution with size \eqn{n}{n} and probability of
 success \eqn{p}{p}.
}
\details{
The Binomial distribution with ... has density ....
}
\note{
Function VaR_binom is a wrapper of the qbinom function from the
stats package.
}
\examples{

E_binom(size = 3, prob = 0.5)

V_binom(size = 3, prob = 0.5)


Etronq_binom(d = 2, size = 3, prob = 0.5)
# Values greater than d
Etronq_binom(d = 0, size = 3, prob = 0.5, less.than.d = FALSE)


VaR_binom(kap = 0.8, size = 5, prob = 0.2)

TVaR_binom(kap = 0.8, size = 5, prob = 0.2)


PGF_binom(t = 1, size = 3, prob = 0.5)

MGF_binom(t = 1, size = 3, prob = 0.5)

}
\concept{Binomial Distribution}
