% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_binom.R
\name{E_binom}
\alias{E_binom}
\title{Espérance d'une loi binomiale
Calcul l'espérance d'une binomiale}
\usage{
E_binom(n, q)
}
\arguments{
\item{n}{nombre d'essais}

\item{q}{probabilité d'un succès indépendant}
}
\description{
Espérance d'une loi binomiale
Calcul l'espérance d'une binomiale
}
