% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_logarithmic.R
\name{E_logarithmique}
\alias{E_logarithmique}
\title{Expected value of the Logarithmic distribution}
\usage{
E_logarithmique(prob)
}
\arguments{
\item{prob}{probability parameter \eqn{\gamma}{gamma}.}
}
\description{
Expected value of the Logarithmic distribution with
 probability parameter \eqn{\gamma}{gamma}.
}
\details{
The Logarithmic distribution with probability parameter \eqn{\gamma}{gam}
 has probability mass function
 \deqn{Pr(X = k) = \frac{-\gamma^{k}}{\ln(1 - \gamma)k}}{-gam^k / (ln(1 - k) k)},
 for \eqn{k = 0, 1, 2, \dots}{k = 0, 1, 2, ...}, and \eqn{\gamma  \in (0, 1)}{0 < gam < 1}]
}
\examples{

E_logarithmique(prob = 0.50)

}
\seealso{
Other Logarithmic Distribution: 
\code{\link{MGF_logarithmic}()},
\code{\link{PGF_logarithmic}()},
\code{\link{V_logarithmique}()}
}
\concept{Logarithmic Distribution}
