% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_unifD.R
\name{E_unifD}
\alias{E_unifD}
\title{Espérance d'une loi uniforme discrète}
\usage{
E_unifD(a, b)
}
\arguments{
\item{a}{borne inférieur}

\item{b}{borne supérieur}
}
\description{
Espérance d'une loi uniforme discrète
}
\details{
loi uniforme discrète
}
