% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_burr.R
\name{Elim_burr}
\alias{Elim_burr}
\title{Espérance limitée d'une loi Burr}
\usage{
Elim_burr(d, alpha, lam, tau)
}
\arguments{
\item{d}{déductible}

\item{alpha}{alpha}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Espérance limitée d'une loi Burr
}
