% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_erlang.R
\name{Elim_erlang}
\alias{Elim_erlang}
\title{Limited mean of the Erlang distribution}
\usage{
Elim_erlang(d, shape, rate = 1/scale, scale = 1/rate)
}
\arguments{
\item{d}{cut-off value.}

\item{shape}{shape parameter \eqn{n}, must be positive integer.}

\item{rate}{\eqn{\beta}{beta} is the rate parameter, must be positive.}

\item{scale}{alternative parameterization to rate parameter, scale = 1 / rate.}
}
\description{
Limited mean of the Erlang distribution
 with shape parameter \eqn{n} and rate parameter \eqn{\beta}{beta}.
}
\details{
The Erlang distribution with shape parameter \eqn{n} and rate parameter
  \eqn{\beta}{b} has density:
  \deqn{f\left(x\right) = \frac{\beta^{n}}{\Gamma(n)} x^{n - 1}%
   \mathrm{e}^{-\beta x}}{f(x) = b^n / \Gamma(n) x^{n - 1} e^{-b x}}
  for \eqn{x \in \mathcal{R}^+}{x > 0}, \eqn{\beta > 0}{b > 0}, \eqn{n \in \mathcal{N}^+}{n = 1, 2, 3, ...}
}
\examples{

# With scale parameter
Elim_erlang(d = 2, shape = 2, scale = 5)

# With rate parameter
Elim_erlang(d = 2, shape = 2, rate = 0.2)

}
\seealso{
Other Erlang Distribution: 
\code{\link{E_erlang}()},
\code{\link{Etronq_erlang}()},
\code{\link{Mexcess_erlang}()},
\code{\link{SL_erlang}()},
\code{\link{TVaR_erlang}()},
\code{\link{V_erlang}()},
\code{\link{VaR_erlang}()},
\code{\link{derlang}()},
\code{\link{kthmoment_erlang}()},
\code{\link{perlang}()}
}
\concept{Erlang Distribution}
