% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elim_pareto.R
\name{Elim_pareto}
\alias{Elim_pareto}
\title{Espérance limitée d'une loi Pareto}
\usage{
Elim_pareto(d, alph, lam)
}
\arguments{
\item{d}{déductible}

\item{alph}{alpha}

\item{lam}{lambda}
}
\description{
Espérance limitée d'une loi Pareto
}
