% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_erlang.R
\name{Etronq_erlang}
\alias{Etronq_erlang}
\title{Espérance tronquée d'une loi Erlang}
\usage{
Etronq_erlang(d, n, b)
}
\arguments{
\item{d}{Deductible}

\item{n}{n}

\item{b}{beta}
}
\description{
Espérance tronquée d'une loi Erlang
}
