% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_erlang.R
\name{Etronq_erlang}
\alias{Etronq_erlang}
\title{Truncated mean of the Erlang distribution}
\usage{
Etronq_erlang(d, shape, rate = 1/scale, scale = 1/rate, less.than.d = TRUE)
}
\arguments{
\item{d}{cut-off value.}

\item{shape}{shape parameter \eqn{n}, must be positive integer.}

\item{rate}{\eqn{\beta}{beta} is the rate parameter, must be positive.}

\item{scale}{alternative parameterization to rate parameter, scale = 1 / rate.}

\item{less.than.d}{logical; if \code{TRUE} (default) truncated mean for values <= d, otherwise, for values > d.}
}
\value{
Function :
 \itemize{
 \item{\code{\link{MGF_erlang}}}{ gives the moment generating function (MGF).}
 \item{\code{\link{derlang}}}{ gives the density function.}
 \item{\code{\link{perlang}}}{ gives the cumulative density function.}
 \item{\code{\link{E_erlang}}}{ gives the expected value.}
 \item{\code{\link{V_erlang}}}{ gives the variance.}
 \item{\code{\link{kthmoment_erlang}}}{ gives the kth moment.}
 \item{\code{\link{Etronq_erlang}}}{ gives the truncated mean.}
 \item{\code{\link{SL_erlang}}}{ gives the stop-loss.}
 \item{\code{\link{Elim_erlang}}}{ gives the limited mean.}
 \item{\code{\link{Mexcess_erlang}}}{ gives the mean excess loss.}
 \item{\code{\link{TVaR_erlang}}}{ gives the Tail Value-at-Risk.}
 \item{\code{\link{VaR_erlang}}}{ gives the Value-at-Risk.}
 }
 Invalid parameter values will return an error detailing which parameter is problematic.
}
\description{
Truncated mean of the Erlang distribution
 with shape parameter \eqn{n} and rate parameter \eqn{\beta}{beta}.
}
\details{
The Erlang distribution with shape parameter \eqn{n} and rate parameter
  \eqn{\beta}{b} has density:
  \deqn{f\left(x\right) = \frac{\beta^{n}}{\Gamma(n)} x^{n - 1}%
   \mathrm{e}^{-\beta x}}{f(x) = b^n / \Gamma(n) x^{n - 1} e^{-b x}}
  for \eqn{x \in \mathcal{R}^+}{x > 0}, \eqn{\beta > 0}{b > 0}, \eqn{n \in \mathcal{N}^+}{n = 1, 2, 3, ...}
}
\examples{

# With scale parameter
Etronq_erlang(d = 2, shape = 2, scale = 5)

# With rate parameter
Etronq_erlang(d = 2, shape = 2, rate = 0.2)

# Values greater than d
Etronq_erlang(d = 2, shape = 2, rate = 0.2, less.than.d = FALSE)

}
\seealso{
Other Erlang Distribution: 
\code{\link{E_erlang}()},
\code{\link{Elim_erlang}()},
\code{\link{MGF_erlang}()},
\code{\link{Mexcess_erlang}()},
\code{\link{SL_erlang}()},
\code{\link{TVaR_erlang}()},
\code{\link{V_erlang}()},
\code{\link{VaR_erlang}()},
\code{\link{derlang}()},
\code{\link{kthmoment_erlang}()},
\code{\link{perlang}()}
}
\concept{Erlang Distribution}
