% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_exp.R
\name{Etronq_exp}
\alias{Etronq_exp}
\title{Espérance tronquée d'une loi exponentielle}
\usage{
Etronq_exp(d, b)
}
\arguments{
\item{d}{montant du déductible}

\item{b}{beta}
}
\description{
Espérance tronquée d'une loi exponentielle
}
