% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_llogis.R
\name{Etronq_llogis}
\alias{Etronq_llogis}
\title{Espérance tronquée d'une loi log-logistique}
\usage{
Etronq_llogis(d, lam, tau)
}
\arguments{
\item{d}{valeur où tronquer}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Espérance tronquée d'une loi log-logistique
}
