% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_llogis.R
\name{Etronq_llogis}
\alias{Etronq_llogis}
\title{Truncated mean of the Loglogistic distribution}
\usage{
Etronq_llogis(d, shape, rate = 1/scale, scale = 1/rate)
}
\arguments{
\item{d}{cut-off value.}

\item{shape}{shape parameter \eqn{\tau}{tau}, must be positive integer.}

\item{rate}{alternative parameterization the scale parameter, rate = 1 / scale.}

\item{scale}{\eqn{\lambda}{lambda} rate parameter, must be positive.}
}
\description{
Truncated mean of the Loglogistic distribution with shape parameter
 \eqn{\tau}{tau} and scale parameter \eqn{\lambda}{lambda}.
}
\details{
The Loglogistic distribution with shape parameter \eqn{\tau}{t} and scale parameter
  \eqn{\lambda}{lam} has density:
  \deqn{\frac{\tau \lambda^\tau x^{\tau -1}}{(\lambda^{\tau }+x^{\tau })^{2}}}{f(x) = t lam^t x^(t - 1) / (lam^t + x^t)^2}
  for \eqn{x \in \mathcal{R}^+}{x > 0}, \eqn{\lambda, \tau > 0}{lam, t > 0}.
}
\examples{

# With scale parameter
Etronq_llogis(d = 2, shape = 2, scale = 5)

# With rate parameter
Etronq_llogis(d = 2, shape = 2, rate = 0.2)

}
\seealso{
Other Loglogistic Distribution: 
\code{\link{Elim_llogis}()},
\code{\link{Mexcess_llogis}()},
\code{\link{SL_llogis}()},
\code{\link{TVaR_llogis}()},
\code{\link{V_llogis}()},
\code{\link{VaR_llogis}()},
\code{\link{kthmoment_llogis}()}
}
\concept{Loglogistic Distribution}
