% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etronq_pois.R
\name{Etronq_pois}
\alias{Etronq_pois}
\title{Truncated mean of the Poisson distribution}
\usage{
Etronq_pois(d, lambda, k0, less.than.d = TRUE)
}
\arguments{
\item{d}{cut-off value.}

\item{lambda}{Rate parameter \eqn{\lambda}{lambda}.}

\item{k0}{point up to which to sum the distribution to approximate the expected value.}

\item{less.than.d}{logical; if \code{TRUE} (default) truncated mean for values <= d, otherwise, for values > d.}
}
\description{
Truncated mean of the Poisson distribution with rate parameter
 \eqn{\lambda}{lambda}.
}
\details{
The Poisson distribution with ... has density ...
}
\examples{

Etronq_pois(d = 0, lambda = 2, k0 = 2E2, less.than.d = FALSE)
Etronq_pois(d = 2, lambda = 2, k0 = 2E2, less.than.d = TRUE)

}
\seealso{
Other Poisson Distribution: 
\code{\link{MGF_pois}()},
\code{\link{PGF_pois}()},
\code{\link{TVaR_pois}()}
}
\concept{Poisson Distribution}
