% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_erlang.R
\name{kthmoment_erlang}
\alias{kthmoment_erlang}
\title{k-ème moment de la loi Erlang}
\usage{
kthmoment_erlang(k, n, b)
}
\arguments{
\item{k}{k-ème moment}

\item{n}{n}

\item{b}{beta}
}
\description{
k-ème moment de la loi Erlang
}
