% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_gamma.R
\name{kthmoment_gamma}
\alias{kthmoment_gamma}
\title{k-ème moment de la loi gamma}
\usage{
kthmoment_gamma(k, shape, rate = 1, scale = 1/rate)
}
\arguments{
\item{k}{k-ème moment}

\item{shape}{alpha}

\item{rate}{(fréquence) beta, 1 par défaut.}

\item{scale}{(échelle) paramétrisation alternative où = 1 / rate.}
}
\description{
k-ème moment de la loi gamma
}
