% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_pareto.R
\name{kthmoment_pareto}
\alias{kthmoment_pareto}
\title{k-ème moment de la loi Pareto}
\usage{
kthmoment_pareto(alpha, lam, k)
}
\arguments{
\item{alpha}{alpha}

\item{lam}{lambda}

\item{k}{k-ème moment, -1 < k < alpha ou alpha > k}
}
\description{
k-ème moment de la loi Pareto
}
