% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_unif.R
\name{kthmoment_unif}
\alias{kthmoment_unif}
\title{kth moment of the Uniform distribution}
\usage{
kthmoment_unif(k, min = 0, max = 1)
}
\arguments{
\item{k}{kth-moment.}

\item{min, max}{lower and upper limits of the distribution. Must be finite.}
}
\description{
kth moment of the Uniform distribution
 with min \eqn{a} and max \eqn{b}.
}
\details{
The (continuous) uniform distribution with min and max parameters \eqn{a}
  and \eqn{b} respectively has density:
  \deqn{f(x) = \frac{1}{b - a} \times \bm{1}_{\{x \in [a, b] \}}}{f(x) = 1 / (b - a) x 1_\{a <= x <= b\}}
  for \eqn{x \in [a, b]}{a <= x <= b}.
}
\examples{

kthmoment_unif(k = 2, min = 3, max = 4)

}
\seealso{
Other Continuous Uniform Distribution: 
\code{\link{E_unif}()},
\code{\link{Elim_unif}()},
\code{\link{Etronq_unif}()},
\code{\link{Mexcess_unif}()},
\code{\link{SL_unif}()},
\code{\link{TVaR_unif}()},
\code{\link{V_unif}()},
\code{\link{VaR_unif}()}
}
\concept{Continuous Uniform Distribution}
