% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthmoment_weibull.R
\name{kthmoment_weibull}
\alias{kthmoment_weibull}
\title{k-ème moment d'une loi Weibull}
\usage{
kthmoment_weibull(k = 1, tau, beta)
}
\arguments{
\item{k}{k-ème moment}

\item{tau}{tau}

\item{beta}{beta}
}
\description{
k-ème moment d'une loi Weibull
}
