% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_shiny.R
\name{lawParametersBox}
\alias{lawParametersBox}
\title{Interactive distribution visualisation (server side)}
\usage{
lawParametersBox(input, output, session, law)
}
\arguments{
\item{input}{input for server side.}

\item{output}{output for server side.}

\item{session}{session for server side.}

\item{law}{Distribution to visualise, one of ...}
}
\description{
Interactive distribution visualisation (server side)
}
\note{
this function should not be used explicitly,
 see \link[Distributacalcul]{Distributacalcul_vis}.
}
