% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_shiny.R
\name{lawParametersBox}
\alias{lawParametersBox}
\title{Interactive distribution visualization (server side)}
\usage{
lawParametersBox(input, output, session, law)
}
\arguments{
\item{input}{input for server side.}

\item{output}{output for server side.}

\item{session}{session for server side.}

\item{law}{Distribution to visualize, one of ...}
}
\value{
Server function for the \code{\link{Distributacalcul_vis}} module.
 Should not be run directly.
}
\description{
Interactive distribution visualization (server side)
}
