% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_shiny.R
\name{lawParametersBoxUI}
\alias{lawParametersBoxUI}
\title{Interactive distribution visualization (UI side)}
\usage{
lawParametersBoxUI(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Interactive distribution visualization (UI side)
}
\note{
this function should not be used explicitly,
 see \link[Distributacalcul]{Distributacalcul_vis}.
}
