% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_shiny.R
\name{lawParametersBoxUI}
\alias{lawParametersBoxUI}
\title{Interactive distribution visualization (UI side)}
\usage{
lawParametersBoxUI(id)
}
\arguments{
\item{id}{id of module}
}
\value{
UI function for the \code{\link{Distributacalcul_vis}} module.
 Should not be run directly.
}
\description{
Interactive distribution visualization (UI side)
}
