% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_burr.R
\name{Mexcess_burr}
\alias{Mexcess_burr}
\title{Mean-Excess loss d'une loi Burr}
\usage{
Mexcess_burr(d, lam, alpha, tau)
}
\arguments{
\item{d}{déductible}

\item{lam}{lambda}

\item{alpha}{alpha}

\item{tau}{tau}
}
\description{
Mean-Excess loss d'une loi Burr
}
