% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_exp.R
\name{Mexcess_exp}
\alias{Mexcess_exp}
\title{Mean Excess-loss d'une loi exponentielle}
\usage{
Mexcess_exp(d, b)
}
\arguments{
\item{d}{montant d de la franchise}

\item{b}{beta}
}
\description{
Mean Excess-loss d'une loi exponentielle
}
