% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mexcess_unif.R
\name{Mexcess_unif}
\alias{Mexcess_unif}
\title{Mean excess loss of the Uniform distribution}
\usage{
Mexcess_unif(d, min = 0, max = 1)
}
\arguments{
\item{d}{cut-off value.}

\item{min, max}{lower and upper limits of the distribution. Must be finite.}
}
\description{
Mean excess loss of the Uniform distribution
 with min \eqn{a} and max \eqn{b}.
}
\details{
The (continuous) uniform distribution with min and max parameters \eqn{a}
  and \eqn{b} respectively has density:
  \deqn{f(x) = \frac{1}{b - a} \times \bm{1}_{\{x \in [a, b] \}}}{f(x) = 1 / (b - a) x 1_\{a <= x <= b\}}
  for \eqn{x \in [a, b]}{a <= x <= b}.
}
\examples{

Mexcess_unif(d = 2, min = 2, max = 4)

}
\seealso{
Other Continuous Uniform Distribution: 
\code{\link{E_unif}()},
\code{\link{Elim_unif}()},
\code{\link{Etronq_unif}()},
\code{\link{SL_unif}()},
\code{\link{TVaR_unif}()},
\code{\link{V_unif}()},
\code{\link{VaR_unif}()},
\code{\link{kthmoment_unif}()}
}
\concept{Continuous Uniform Distribution}
