% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_BINCOMP.R
\name{p_BINCOMP}
\alias{p_BINCOMP}
\title{Fonction de répartition d'une loi binomiale composée}
\usage{
p_BINCOMP(x, n, q, shape, rate, ko, distr_severity = "Gamma")
}
\arguments{
\item{x}{x}

\item{n}{n pour la binomiale}

\item{q}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{ko}{borne pour la somme de la fonction de répartition}

\item{distr_severity}{choix de distribution de sévérité.}
}
\description{
Fonction de répartition d'une loi binomiale composée
}
