% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_BNCOMP.R
\name{p_BNCOMP}
\alias{p_BNCOMP}
\title{Répartition d'une loi binomiale négative composée}
\usage{
p_BNCOMP(x, size, prob, shape, rate, ko, distr_severity = "Gamma")
}
\arguments{
\item{x}{x}

\item{size}{r pour la binomiale négative (nombre d'échecs).}

\item{prob}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{ko}{borne pour la sommation de la fonction de répartition}

\item{distr_severity}{choix de distribution de sévérité.}
}
\description{
Répartition d'une loi binomiale négative composée
}
\examples{
p_BNCOMP(x = 5E5, ko = 1E3, size = 5, prob = 1/5, shape = 2, rate = 1/5000, distr_severity = "Gamma")

}
