% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_PCOMP.R
\name{p_PCOMP}
\alias{p_PCOMP}
\title{Fonction de répartition d'une loi poisson composée}
\usage{
p_PCOMP(x, lambda, shape, rate, ko = 300, distr_severity = "Gamma")
}
\arguments{
\item{x}{x}

\item{lambda}{paramètre lambda pour la poisson}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{ko}{borne pour la sommation de la fonction de répartition}

\item{distr_severity}{choix de distribution de sévérité.}
}
\description{
Fonction de répartition d'une loi poisson composée
}
