% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perlang.R
\name{perlang}
\alias{perlang}
\title{Fonction de répartition de la loi Erlang}
\usage{
perlang(x, n, b, lower.tail = T)
}
\arguments{
\item{x}{x}

\item{n}{n}

\item{b}{beta}

\item{lower.tail}{Si vrai (défaut) alors fonction de répartition. Si faux, de survie.}
}
\description{
Fonction de répartition de la loi Erlang
}
