% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGF_logarithmic.R
\name{PGF_logarithmic}
\alias{PGF_logarithmic}
\title{Probability Generating Function of the Logarithmic distribution}
\usage{
PGF_logarithmic(t, prob)
}
\arguments{
\item{t}{t.}

\item{prob}{probability parameter \eqn{\gamma}{gamma}.}
}
\description{
Probability Generating Function (PGF) of the Logarithmic
 distribution with probability parameter \eqn{\gamma}{gamma}.
}
\details{

}
\examples{

PGF_logarithmic(t = .2, prob = 0.50)

}
\seealso{
Other Logarithmic Distribution: 
\code{\link{E_logarithmique}()},
\code{\link{MGF_logarithmic}()},
\code{\link{V_logarithmique}()}
}
\concept{Logarithmic Distribution}
