% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punifD.R
\name{punifD}
\alias{punifD}
\title{Répartition d'une loi uniforme discrète}
\usage{
punifD(q, a, b)
}
\arguments{
\item{q}{quantile}

\item{a}{borne inférieur}

\item{b}{borne supérieur}
}
\description{
Répartition d'une loi uniforme discrète
}
\details{
loi uniforme discrète
}
