% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_erlang.R
\name{SL_erlang}
\alias{SL_erlang}
\title{Stop-Loss of the Erlang Distribution}
\usage{
SL_erlang(d, shape, scale, rate = 1/scale)
}
\arguments{
\item{shape}{shape parameter \eqn{n}, must be positive integer.}

\item{scale}{\eqn{\beta}{beta} is the scale parameter, must be positive.}

\item{rate}{alternative parameterization to scale parameter, rate = 1 / scale.}
}
\description{
Stop-loss of the Erlang distribution with shape
 parameter \eqn{n} and rate parameter \eqn{\beta}{b}.
}
\examples{

# With scale parameter
SL_erlang(d = 2, shape = 2, scale = 5)

# With rate parameter
SL_erlang(d = 2, shape = 2, rate = 0.2)

}
\seealso{
[derlang()] for the probability mass function,
 [perlang()] for the cumulative probability mass function,
 [V_erlang()] for the variance,
 [Mexcess_erlang()] for the Mean Excess Loss,
 [Elim_erlang()] for the limited expected value,
 [Etronq_erlang()] for the truncated mean,
 [E_erlang()] for the expected value, and
 [kthmoment_erlang()] for the \eqn{k}th moment.
}
