% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_norm.R
\name{SL_norm}
\alias{SL_norm}
\title{Stop-Loss de la loi normale}
\usage{
SL_norm(d, mu = 0, sig = 1)
}
\arguments{
\item{d}{déductible}

\item{mu}{moyenne mu}

\item{sig}{sigma}
}
\description{
Stop-Loss de la loi normale
}
