% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_BINCOMP.R
\name{TVaR_BINCOMP}
\alias{TVaR_BINCOMP}
\title{Tail Value-at-Risk d'une loi binomiale composée}
\usage{
TVaR_BINCOMP(kap, size, prob, shape, rate, vark, ko, distr_severity = "Gamma")
}
\arguments{
\item{kap}{Niveau de confiance désiré}

\item{size}{n pour la binomiale}

\item{prob}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{vark}{Value at Risk (VaR) calculé au même niveau de confiance kap}

\item{ko}{borne pour la somme dans le calcul de la TVaR}

\item{distr_severity}{choix de distribution de sévérité}
}
\description{
Tail Value-at-Risk d'une loi binomiale composée
}
\details{
Cette formule nécessite la formule de la VaR_bincomp (déjà installée avec le package tvarPackage). Cette formule utilise aussi un choix de (présentement) 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
