% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_BNCOMP.R
\name{TVaR_BNCOMP}
\alias{TVaR_BNCOMP}
\title{Tail Value-at-risk d'une loi binomiale négative composée}
\usage{
TVaR_BNCOMP(kappa, r, q, shape, rate, vark, ko, distr_severity = "Gamma")
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{r}{r pour la binomiale}

\item{q}{probabilité q pour la binomiale négative}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{vark}{Value at Risk (VaR) calculé au même niveau de confiance kappa}

\item{ko}{borne pour la somme dans le calcul de la TVaR.}

\item{distr_severity}{choix de distribution de sévérité.}
}
\description{
Tail Value-at-risk d'une loi binomiale négative composée
}
\details{
Cette formule nécessite la formule de la VaR_bncomp (déjà installée avec le package tvarPackage). Cette formule utilise aussi un choix de (présentement) 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
