% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_erlang.R
\name{TVaR_erlang}
\alias{TVaR_erlang}
\title{Tail Value-at-risk d'une loi Erlang}
\usage{
TVaR_erlang(kappa, n, b, vark)
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{n}{n}

\item{b}{beta}

\item{vark}{Value at Risk (VaR) calculé au même niveau de confiance kappa}
}
\description{
Tail Value-at-risk d'une loi Erlang
}
\details{
Cette formule nécessite la formule de la VaR_erlang (déjà installée avec le package tvarPackage)
}
