% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_llogis.R
\name{TVaR_llogis}
\alias{TVaR_llogis}
\title{Tail Value-at-risk d'une loi log-logistique}
\usage{
TVaR_llogis(kappa, lam, tau)
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{lam}{lam}

\item{tau}{tau}
}
\description{
Tail Value-at-risk d'une loi log-logistique
}
