% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_norm.R
\name{TVaR_norm}
\alias{TVaR_norm}
\title{Tail Value-at-Risk d'une loi normale}
\usage{
TVaR_norm(kappa, mu, sig)
}
\arguments{
\item{kappa}{pourcentage de confiance désiré}

\item{mu}{mu}

\item{sig}{sigma}
}
\description{
Tail Value-at-Risk d'une loi normale
}
