% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_PCOMP.R
\name{TVaR_PCOMP}
\alias{TVaR_PCOMP}
\title{Tail Value-at-risk d'une loi poisson composée}
\usage{
TVaR_PCOMP(kappa, lambda, shape, rate, vark, ko = 300,
  distr_severity = "Gamma")
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{lambda}{lambda}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{vark}{Value at Risk (VaR) calculé au même niveau de confiance kappa}

\item{ko}{borne pour la somme dans le calcul de la TVaR}

\item{distr_severity}{choix de distribution de sévérité.}
}
\description{
Tail Value-at-risk d'une loi poisson composée
}
\details{
Cette formule nécessite la formule de la VaR_pcomp (déjà installée avec le package tvarPackage). Cette formule utilise aussi un choix de (présentement) 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
