% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_pois.R
\name{TVaR_pois}
\alias{TVaR_pois}
\title{Tail Value-at-risk de la loi Poisson}
\usage{
TVaR_pois(kappa, lam)
}
\arguments{
\item{kappa}{niveau de confiance désiré}

\item{lam}{lambda, nombre d'occurences moyen dans un laps de temps donné.}
}
\description{
Tail Value-at-risk de la loi Poisson
}
