% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_pois.R
\name{TVaR_pois}
\alias{TVaR_pois}
\title{Tail Value-at-risk of the Poisson distribution}
\usage{
TVaR_pois(kappa, lambda, k0)
}
\arguments{
\item{lambda}{Rate parameter \eqn{\lambda}{lambda}.}
}
\description{
Tail Value-at-Risk of the Poisson distribution with rate parameter
 \eqn{\lambda}{lambda}.
}
\examples{

TVaR_pois(kappa = 0.8, lambda = 3, k0 = 2E2)

}
