% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVaR_unif.R
\name{TVaR_unif}
\alias{TVaR_unif}
\title{Tail Value-at-Risk of the Uniform distribution}
\usage{
TVaR_unif(kap, min, max)
}
\arguments{
\item{kap}{probability.}

\item{min}{minimum a}

\item{max}{maximum b}
}
\description{
Tail Value-at-Risk of the Uniform distribution with shape
 parameters \eqn{a}{a} and \eqn{b}{b}.
}
\examples{

# With scale parameter
TVaR_unif(kap = .2, min = 3, max = 4)

}
