% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_binom.R
\name{V_binom}
\alias{V_binom}
\title{Variance d'une loi binomiale}
\usage{
V_binom(n, q)
}
\arguments{
\item{n}{nombre d'essais}

\item{q}{probabilité d'un succès indépendant}
}
\description{
Variance d'une loi binomiale
}
