% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_IG.R
\name{V_IG}
\alias{V_IG}
\title{Variance de la loi inverse gaussienne}
\usage{
V_IG(mu, beta = dispersion * mu^2, dispersion = beta/mu^2)
}
\arguments{
\item{mu}{mu}

\item{beta}{beta = dispersion * mu^2}

\item{dispersion}{dispersion = beta / mu^2}
}
\description{
Variance de la loi inverse gaussienne
}
