% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_logarithmic.R
\name{V_logarithmique}
\alias{V_logarithmique}
\title{Variance of the Logarithmic distribution}
\usage{
V_logarithmique(prob)
}
\arguments{
\item{prob}{probability parameter \eqn{\gamma}{gamma}.}
}
\description{
Variance of the Logarithmic distribution with
 probability parameter \eqn{\gamma}{gamma}.
}
\details{

}
\examples{

V_logarithmique(prob = 0.50)

}
\seealso{
Other Logarithmic Distribution: 
\code{\link{E_logarithmique}()},
\code{\link{MGF_logarithmic}()},
\code{\link{PGF_logarithmic}()}
}
\concept{Logarithmic Distribution}
