% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_logarithmique.R
\name{V_logarithmique}
\alias{V_logarithmique}
\title{Variance d'une loi logarithmique}
\usage{
V_logarithmique(gam)
}
\arguments{
\item{gam}{paramètre gamma}
}
\description{
Variance d'une loi logarithmique
}
