% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_negbinom.R
\name{V_negbinom}
\alias{V_negbinom}
\title{Variance d'une loi binomiale négative}
\usage{
V_negbinom(r, p = (1/(1 + beta)), beta = ((1 - p)/p), nb_tries = F)
}
\arguments{
\item{r}{Number of successful trials.}

\item{p}{Probability of success.}

\item{beta}{Alternative parameterization of the negative binomial
distribution where beta = (1 - p) / p.}

\item{nb_tries}{logical; if \code{FALSE} (default) number of trials
until the \code{r}th success, otherwise, number of failures until
the \code{r}th success.}
}
\description{
Variance of the negative binomial distribution with
 parameters \eqn{r} (number of successful trials) and \eqn{p}
 (probability of success).
}
\details{
When \eqn{k} is the number of failures until the \eqn{r}th success,
  with a probability \eqn{p} of a success, the negative binomial has density:
  \deqn{\binom{r + k - 1}{k} (p)^{r} (1 - p)^{k}}{%
   Pr(M = k) = choose(r + k - 1, k) p^r (1 - p)^k}
  for \eqn{k \in \{0, 1, \dots \}}{k = 0, 1, 2, ...}

 When \eqn{k} is the number of trials until the \eqn{r}th success,
  with a probability \eqn{p} of a success, the negative binomial has density:
  \deqn{\binom{k - 1}{r - 1} (p)^{r} (1 - p)^{k - r}}{%
   Pr(M = k) = choose(k - 1, r - 1) p^r (1 - p)^(k - r)}
  for \eqn{k \in \{r, r + 1, r + 2, \dots \}}{k = r, r + 1, r + 2, ...}

 The alternative parameterization of the negative binomial with parameter
  \eqn{\beta}{beta}, and \eqn{k} being the number of trials, has density:
  \deqn{\frac{\Gamma(r + k)}{\Gamma(r) k!} \left(\frac{1}{1 + \beta}\right)^{r}%
   \left(\frac{\beta}{1 +  \beta}\right)^{k - r}}{Pr(M = k) = %
   \Gamma(r + k)/(\Gamma(r) k!) (1/(1 + beta))^r (beta/(1 + beta))^(k - r)}
  for \eqn{k \in \{0, 1, \dots \}}{k = 0, 1, 2, ...}

The variance of the negative binomial where \eqn{p} is the
 probability of success and:
 \eqn{k} is the number of failures or the number of trials until the
  \eqn{r}th success is:
  \deqn{\text{V}(M) = \frac{r(1 - p)}{p^2}}{E(M) = r(1 - p)/(p^2)}

 \eqn{k} is the number of trials until the \eqn{r}th success is:
  \deqn{\text{V}(X) = r \beta (1 + \beta)}{E(M) = r beta (1 + beta)}
}
\examples{

# Where k is the number of trials for a rth success
V_negbinom(r = 2, p = .4)

# Where k is the number of failures before a rth success
V_negbinom(r = 2, p = .4, nb_tries = T)

# With alternative parameterization where k is the number of trials
V_negbinom(r = 2, beta = 1.5)

}
\seealso{
[d_negbinom()] for the probability mass function, [p_negbinom()]
 for the cumulative probability mass function, and [E_negbinom()]
 for the mean.

Other Negative Binomial Distribution: \code{\link{E_negbinom}},
  \code{\link{d_negbinom}}, \code{\link{p_negbinom}}
}
\concept{Negative Binomial Distribution}
