% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_unif.R
\name{V_unif}
\alias{V_unif}
\title{Variance d'une loi uniforme
Calcul la variance d'une uniforme}
\usage{
V_unif(a, b)
}
\arguments{
\item{a}{minimum}

\item{b}{maximum}
}
\description{
Variance d'une loi uniforme
Calcul la variance d'une uniforme
}
