% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_unifD.R
\name{V_unifD}
\alias{V_unifD}
\title{Variance d'une loi uniforme discrète}
\usage{
V_unifD(a, b)
}
\arguments{
\item{a}{borne inférieur}

\item{b}{borne supérieur}
}
\description{
Variance d'une loi uniforme discrète
}
\details{
loi uniforme discrète
}
