% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_beta.R
\name{VaR_beta}
\alias{VaR_beta}
\title{Value-at-Risk  d'une loi Beta}
\usage{
VaR_beta(k, a, b)
}
\arguments{
\item{k}{kappa (doit être entre 0 et 1)}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Value-at-Risk  d'une loi Beta
}
