% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_BINCOMP.R
\name{VaR_BINCOMP}
\alias{VaR_BINCOMP}
\title{Value-at-risk of the compound Binomiale distribution}
\usage{
VaR_BINCOMP(kap, n, q, shape, rate, ko = 300, distr_severity = "Gamma")
}
\arguments{
\item{kap}{Niveau de confiance désiré}

\item{n}{n pour la binomiale}

\item{q}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{ko}{borne pour l'optimisation numérique de la VaR}

\item{distr_severity}{choix de distribution de sévérité. Gamma uniquement pour l'instant.}
}
\description{
Value-at-risk of the compound Binomiale distribution
}
