% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_BINCOMP.R
\name{VaR_BINCOMP}
\alias{VaR_BINCOMP}
\title{Value-at-risk d'une loi binomiale composée}
\usage{
VaR_BINCOMP(k, n, q, shape, rate, ko = 300, distr_severity = "Gamma")
}
\arguments{
\item{k}{Niveau de confiance désiré}

\item{n}{n pour la binomiale}

\item{q}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{ko}{borne pour l'optimisation numérique de la VaR}

\item{distr_severity}{choix de distribution de sévérité. Gamma ou Lognormale}
}
\description{
Value-at-risk d'une loi binomiale composée
}
\details{
Cette formule utilise aussi un choix de 2 distributions pour la sévérité; soit la Gamma (par défaut) ou la Lognormale.
}
