% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_binom.R
\name{VaR_binom}
\alias{VaR_binom}
\title{Value-at-risk of the Binomial distribution}
\usage{
VaR_binom(kappa, size, prob)
}
\arguments{
\item{size}{Number of trials (0 or more).}

\item{prob}{Probability of success on each trial.}
}
\description{
Value-at-risk of the Binomial distribution with size
 \eqn{n}{n} and probability of success \eqn{p}{p}. Wrapper of qbinom.
}
\examples{

VaR_binom(kappa = 0.8, size = 5, prob = 0.2)

}
