% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_binom.R
\name{VaR_binom}
\alias{VaR_binom}
\title{Value-at-Risk of the Binomial distribution}
\usage{
VaR_binom(kap, size, prob)
}
\arguments{
\item{kap}{probability.}

\item{size}{Number of trials (0 or more).}

\item{prob}{Probability of success on each trial.}
}
\description{
Value-at-Risk of the Binomial distribution with size
 \eqn{n}{n} and probability of success \eqn{p}{p}.
}
\details{
The Binomial distribution
}
\note{
Function VaR_binom is a wrapper for the qbinom function from the
stats package.
}
\examples{

VaR_binom(kap = 0.8, size = 5, prob = 0.2)

}
\seealso{
Other Binomial Distribution: 
\code{\link{E_binom}()},
\code{\link{Etronq_binom}()},
\code{\link{MGF_binom}()},
\code{\link{PGF_binom}()},
\code{\link{TVaR_binom}()},
\code{\link{V_binom}()}
}
\concept{Binomial Distribution}
