% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_BNCOMP.R
\name{VaR_BNCOMP}
\alias{VaR_BNCOMP}
\title{Value-at-risk d'une loi binomiale négative composée}
\usage{
VaR_BNCOMP(kappa, r, q, shape, rate, ko = 300,
  distr_severity = "Gamma")
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{r}{r pour la binomiale négative (nombre d'échecs).}

\item{q}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{ko}{borne pour l'optimisation numérique de la VaR}

\item{distr_severity}{choix de distribution de sévérité. Gamma ou Lognormale}
}
\description{
Value-at-risk d'une loi binomiale négative composée
}
\details{
Cette formule utilise aussi un choix de 2 distributions pour la sévérité; soit la Gamma (par défaut) ou la Lognormale.
}
