% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_BNCOMP.R
\name{VaR_BNCOMP}
\alias{VaR_BNCOMP}
\title{Value-at-risk d'une loi binomiale négative composée}
\usage{
VaR_BNCOMP(kap, size, prob, shape, rate, distr_severity = "Gamma", ko)
}
\arguments{
\item{kap}{Niveau de confiance désiré}

\item{size}{r pour la binomiale négative}

\item{prob}{probabilité q pour la binomiale}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{distr_severity}{choice of severity distribution Gamma (default) or Lognormal}

\item{ko}{borne pour la somme dans le calcul de la TVaR.}
}
\description{
Value-at-risk d'une loi binomiale négative composée
}
\examples{

VaR_BNCOMP(kap = 0.99, size = 4, prob = 0.4, shape = 5, rate = 10, distr_severity = "Gamma", ko = 1500)

}
