% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_burr.R
\name{VaR_burr}
\alias{VaR_burr}
\title{Value-at-risk de la loi Burr}
\usage{
VaR_burr(k, alpha, lam, tau)
}
\arguments{
\item{k}{niveau de confiance désiré}

\item{alpha}{alpha}

\item{lam}{lambda}

\item{tau}{tau}
}
\description{
Value-at-risk de la loi Burr
}
