% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_exp.R
\name{VaR_exp}
\alias{VaR_exp}
\title{Value-at-risk of the Exponential distribution}
\usage{
VaR_exp(kap, rate = 1/scale, scale = 1/rate)
}
\arguments{
\item{kap}{pourcentage de confiance désiré.}

\item{rate}{beta.}

\item{scale}{alternate parametrization, scale = 1 / rate.}
}
\description{
Value-at-Risk of the Exponential distribution with rate
  parameter \eqn{\beta}{beta}.
}
\examples{

# With scale parameter
VaR_exp(kap = .2, shape = 3, scale = 4)

# With rate parameter
VaR_exp(kap = .2, shape = 3, rate = 0.25)

}
