% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_FGEN.R
\name{VaR_FGEN}
\alias{VaR_FGEN}
\title{VaR de la F-Généralisé}
\usage{
VaR_FGEN(x, kappa, lambda, alpha, tau)
}
\arguments{
\item{x}{x}

\item{kappa}{niveau de confiance désiré}

\item{lambda}{lambda, > 0}

\item{alpha}{alpha, >0}

\item{tau}{tau, >0}
}
\description{
VaR de la F-Généralisé
}
