% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_gamma.R
\name{VaR_gamma}
\alias{VaR_gamma}
\title{Value-at-risk d'une loi gamma}
\usage{
VaR_gamma(kappa, a, b)
}
\arguments{
\item{kappa}{Niveau de confiance désiré}

\item{a}{alpha}

\item{b}{beta}
}
\description{
Value-at-risk d'une loi gamma
}
\details{
Pour trouver la VaR d'une gamme, on peut aussi procéder par optimisation numérique (`uniroot()` ou `optimize()`)
}
