% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_pareto.R
\name{VaR_pareto}
\alias{VaR_pareto}
\title{Value-at-risk of the Pareto distribution}
\usage{
VaR_pareto(kap, shape, rate)
}
\arguments{
\item{kap}{kap}

\item{shape}{alpha}

\item{rate}{lambda}
}
\description{
Value-at-risk of the Pareto distribution
}
