% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_PCOMP.R
\name{VaR_PCOMP}
\alias{VaR_PCOMP}
\title{Fonction de répartition d'une loi poisson composée}
\usage{
VaR_PCOMP(kappa, lambda, shape, rate, ko = 300,
  distr_severity = "Gamma")
}
\arguments{
\item{kappa}{niveau de confiance désiré}

\item{lambda}{paramètre lambda pour la poisson}

\item{shape}{alpha pour la Gamma et mu pour la lognormale}

\item{rate}{beta pour la Gamma et sigma^2 pour la lognormale}

\item{ko}{borne pour l'optimisation de la VaR}

\item{distr_severity}{choix de distribution de sévérité. Gamma (par défaut) ou Lognormale}
}
\description{
Fonction de répartition d'une loi poisson composée
}
\details{
Cette formule utilise aussi un choix de 2 distributions pour la sévérité; soit la Gamma ou la Lognormale.
}
