% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR_PCOMP.R
\name{VaR_PCOMP}
\alias{VaR_PCOMP}
\title{Value-at-Risk d'une loi poisson composée}
\usage{
VaR_PCOMP(kap, lambda, shape, rate, distr_severity = "Gamma", ko)
}
\arguments{
\item{kap}{niveau de confiance désiré}

\item{lambda}{paramètre lambda pour la poisson}

\item{shape}{alpha pour la Gamma}

\item{rate}{beta pour la Gamma}

\item{distr_severity}{choix de distribution de sévérité.}

\item{ko}{borne pour l'optimisation de la VaR}
}
\description{
Value-at-Risk d'une loi poisson composée
}
